refactor(trip-detail): extract tabs/header/map band out of the god-screen
Pure refactor, zero visual and zero behavior change: lifts the itinerary
tab, bookings tab, header stack, and map band out of
trip_detail_screen.dart (8,338 -> 4,446 lines) so wave 2 can run the
itinerary/bookings/header/map redesign lanes in parallel on separate
files.

- itinerary_tab.dart + bookings_tab.dart: part files (verbatim moves via
  extensions on the screen's State; same library, so private members and
  imports are untouched). One mechanical touch: extension bodies can't
  call the protected State.setState, so 11 sites call the screen's
  one-line _rebuild pass-through; two sites qualify the static
  _groundModeOf by type name.
- trip_header_card.dart: TripHeaderCard public widget (title/meta block,
  Next Step card, Continue-chat card, _OverviewText). Screen state
  arrives as params, actions as callbacks; the next-step session flags
  move into the widget's State.
- map_band.dart: TripDetailMapBand public widget (the rounded map card),
  tripDetailMapBandSliver (the pinned-wide/scroll-away slot), and the two
  band heights as public constants. The pinned delegate stays the
  screen's (it also pins the tab row) and arrives as a builder.

Proof: flutter analyze clean; full suite green with zero test edits
(1701/1701, literal 'All tests passed!'); 13/13 before/after frames
byte-identical on the seeded Big Summer Adventure trip (3 tabs + header +
map band, wide + narrow + scroll stops).
