+++ src/packages/flutter-app/lib/screens/trip_detail_screen.dart
@@ -24,7 +24,6 @@ import '../providers/trips_provider.dart';
 import '../providers/recent_trip_provider.dart';
 import '../utils/errors.dart';
 import '../providers/booking_todos_provider.dart';
-import '../providers/home_overlay_provider.dart';
 import '../providers/preferences_provider.dart';
 import '../providers/api_client_provider.dart';
 import '../providers/plan_provider.dart';
@@ -46,7 +45,6 @@ import '../services/api_client.dart' show isTransientError;
 import '../services/trip_cache.dart';
 import '../services/trips_api_service.dart' show TripEndpointsException;
 import '../theme/app_colors.dart';
-import '../theme/app_typography.dart';
 import '../theme/spacing.dart';
 import '../utils/calendar_links.dart';
 import '../utils/clothing_recs.dart';
@@ -62,7 +60,6 @@ import '../utils/trip_format.dart';
 import '../utils/trip_legs.dart';
 import '../widgets/add_itinerary_item_dialog.dart';
 import '../widgets/add_to_trip_sheet.dart';
-import '../widgets/app_map.dart';
 import '../services/api_client.dart' show ApiException;
 import '../widgets/booked_expense_prompt.dart';
 import '../widgets/booking_detail_row.dart';
@@ -77,24 +74,24 @@ import '../widgets/empty_state.dart';
 import '../widgets/city_events_sheet.dart';
 import '../widgets/hover_reveal.dart';
 import '../widgets/local_rec_card.dart';
-import '../widgets/map_leg_chips.dart';
-import '../widgets/next_step_card.dart';
 import '../widgets/offline_banner.dart';
 import '../widgets/place_photo_card.dart';
-import '../widgets/plan_progress_sheet.dart';
 import '../widgets/source_links_card.dart';
 import '../widgets/status_pill.dart';
 import '../widgets/trip_actions_sheet.dart';
 import '../widgets/trip_airports_sheet.dart';
 import '../widgets/trip_details_dialog.dart';
-import '../widgets/trip_map.dart';
 import '../widgets/trip_refine_panel.dart';
 import '../widgets/wear_pack_sheet.dart';
 import 'flight_search_screen.dart';
 import 'local_guide_detail_screen.dart';
 import 'trip_detail_derivation.dart';
 import 'trip_map_screen.dart';
+import '../widgets/trip_detail/trip_header_card.dart';
+import '../widgets/trip_detail/map_band.dart';
 import '../utils/snack.dart';
+part '../widgets/trip_detail/itinerary_tab.dart';
+part '../widgets/trip_detail/bookings_tab.dart';
 
 /// A geographic coordinate used to resolve an itinerary place to its nearest
 /// bookable airport when the place name has no IATA match.
@@ -149,6 +146,12 @@ class TripDetailScreen extends ConsumerStatefulWidget {
 
 class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     with WidgetsBindingObserver {
+  /// setState pass-through for the part-file extensions (specs/
+  /// trip-detail-extract): [State.setState] is protected, so extension
+  /// members can't call it directly. Identical semantics — same object,
+  /// same closure, same rebuild.
+  void _rebuild(VoidCallback fn) => setState(fn);
+
   Trip? _trip;
   bool _loading = true;
   // The raw caught load error (not a string): the error screen classifies it
@@ -170,13 +173,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   Object? _chatError;
   bool _chatResumeTried = false;
 
-  // Next Step CTA session state (specs/next-step-cta): the all_set
-  // celebration shows only after this session actually watched a step resolve
-  // (a trip opened already-complete shows nothing), and its X hides it for
-  // the rest of the session. Deliberately not persisted — steps carry no
-  // stable identity to key a durable dismissal on.
-  bool _hadNextStep = false;
-  bool _allSetDismissed = false;
   // The active view state. 'all' renders the grouped itinerary; 'bookings'
   // and 'unbooked' are the Bookings view (entered from the header tab) and
   // its left-to-book scope (the chip inside that view) — both replace the
@@ -645,31 +641,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   // ([_dateChipWidth]), so calendar icons, range starts, and nights suffixes
   // form columns across rows.
 
-  /// Calendar icon size and trailing gap inside the chip — consumed by both
-  /// the render and the measurement ([_chipIconSpan]) so they cannot drift.
-  static const double _chipIconSize = 14;
-  static const double _chipIconGap = 4;
-  static const double _chipIconSpan = _chipIconSize + _chipIconGap;
-
-  /// Minimum gap between the range text and the nights suffix. Also the
-  /// error budget for TextPainter-vs-Text measurement divergence: a chip can
-  /// only spuriously ellipsize if measurement undershoots by more than this.
-  static const double _chipInnerGap = 8;
-
-  /// Pathological ceiling (kept from the pre-column layout): real localized
-  /// chips measure well under it; absurd text scales ellipsize instead of
-  /// overflowing the header row. Never scale this with the textScaler — the
-  /// chip is a rigid child of the header row and a scaled cap could exceed a
-  /// phone row's width.
-  static const double _chipMaxWidth = 200;
-
-  /// Pinned-header heights, shared by the build method and the Today scroll
-  /// math so the two can never drift apart.
-  static const double _mapHeaderHeight =
-      12 + 340 + 12; // top gap + map + bottom gap
-  /// Map card height on phones, where the map scrolls away instead of
-  /// pinning (a preview — the full-screen map is one tap away).
-  static const double _mapHeightNarrow = 180;
   // Itinerary/Bookings/Budget tab row (44 — real touch targets for the view
   // tabs)
   // + bottom padding (8) + breathing room (4); the header is one
@@ -685,7 +656,7 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   /// (`maxScrollObstructionExtentBefore`).
   double _pinnedChrome(Trip trip) {
     final mapShown = _derive(trip).mapShown;
-    return ((_mapPinned && mapShown) ? _mapHeaderHeight : 0) +
+    return ((_mapPinned && mapShown) ? mapBandHeaderHeight : 0) +
         _listHeaderHeight;
   }
 
@@ -1428,483 +1399,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   // The per-city booking-slot partition lives in
   // [TripDerivation.groupedBookings] (trip_detail_derivation.dart).
 
-  /// The one "Booked" writer: flips the todo and (when the slot has one) the
-  /// matched confirmed record in lockstep, so every reader of either flag —
-  /// checklist rows, calendar/.ics export, print packet — agrees with the
-  /// single visible checkbox. Optimistic on all touched lists; any failure
-  /// rolls every list back (a partial server success self-heals on the next
-  /// toggle or sync).
-  Future<void> _setRowBooked(bool booked,
-      {BookingTodo? todo, Accommodation? stay, TripSegment? segment}) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final prevTodos = _bookingTodos;
-    final prevStays = _stays;
-    final prevSegments = _segments;
-    setState(() {
-      if (todo != null) {
-        _bookingTodos = [
-          for (final t in _bookingTodos)
-            if (t.id == todo.id) t.copyWith(booked: booked) else t,
-        ];
-      }
-      if (stay != null) {
-        _stays = [
-          for (final a in _stays)
-            if (a.id == stay.id) a.copyWith(booked: booked) else a,
-        ];
-      }
-      if (segment != null) {
-        _segments = [
-          for (final s in _segments)
-            if (s.id == segment.id) s.copyWith(booked: booked) else s,
-        ];
-      }
-    });
-    try {
-      await Future.wait<void>([
-        if (todo != null)
-          ref
-              .read(bookingTodosApiServiceProvider)
-              .setBooked(widget.tripId, todo.id, booked),
-        if (stay != null)
-          ref
-              .read(accommodationsApiServiceProvider)
-              .update(widget.tripId, stay.id, {'booked': booked}),
-        if (segment != null)
-          ref
-              .read(transportApiServiceProvider)
-              .updateSegment(widget.tripId, segment.id, {'booked': booked}),
-      ]);
-    } catch (e) {
-      if (mounted) {
-        setState(() {
-          _bookingTodos = prevTodos;
-          _stays = prevStays;
-          _segments = prevSegments;
-        });
-      }
-      _showSnack(l10n.tripUpdateFailed(friendlyError(l10n, e)));
-      return;
-    }
-    // The booked flip IS the Next Step card's advance signal: phase 3 walks
-    // the booking slots and phase 5 aggregates them, so both read the flag
-    // this call just wrote (specs/next-step-cta). Only after the server
-    // accepted it — a rolled-back optimistic flip must not move the card.
-    if (mounted) _invalidateReview();
-
-    // Budget autopopulate rides the flip only AFTER the server accepted it
-    // (a rolled-back optimistic flip must never create or delete money).
-    if (!mounted) return;
-    if (booked) {
-      await _maybePromptBudgetExpense(todo: todo, stay: stay, segment: segment);
-    } else {
-      await _removeLinkedAutoExpense([
-        if (todo != null) todo.id,
-        if (stay != null) stay.id,
-        if (segment != null) segment.id,
-      ]);
-    }
-  }
-
-  /// Booked-flip budget autopopulate (specs/budget-v2): right after a
-  /// false→true flip lands, offer to record the price — the moment the
-  /// traveler actually has it in hand. Dedupe first: a linked expense for
-  /// ANY of the flip's row ids (todo + matched stay/segment flip together)
-  /// means this booking is already in the budget — a re-book after a manual
-  /// takeover stays silent; the server's upsert-by-source is the backstop
-  /// when the read fails. The link rides the most durable row
-  /// (stay ?? segment ?? todo). Undo on the confirmation snackbar deletes
-  /// the expense it just created — the booked state itself stays (the
-  /// checkbox is one tap away; Undo here is about the money).
-  Future<void> _maybePromptBudgetExpense(
-      {BookingTodo? todo, Accommodation? stay, TripSegment? segment}) async {
-    if (!mounted || _readOnly) return;
-    final l10n = context.l10n;
-    var currency = 'USD';
-    try {
-      currency =
-          (await ref.read(budgetProvider(widget.tripId).future)).currency;
-    } catch (_) {} // prompt still works; USD is the budget default
-    final ids = {
-      if (todo != null) todo.id,
-      if (stay != null) stay.id,
-      if (segment != null) segment.id,
-    };
-    try {
-      final expenses = await ref.read(expensesProvider(widget.tripId).future);
-      if (expenses.any((e) => ids.contains(e.sourceId))) return;
-    } catch (_) {}
-    if (!mounted) return;
-    final draft = await showBookedExpensePrompt(
-      context,
-      currency: currency,
-      prefill:
-          deriveBookedExpensePrefill(todo: todo, stay: stay, segment: segment),
-    );
-    if (draft == null || !mounted) return;
-    final source = stay != null
-        ? (kind: 'accommodation', id: stay.id)
-        : segment != null
-            ? (kind: 'segment', id: segment.id)
-            : (kind: 'booking_todo', id: todo!.id);
-    try {
-      final added = await ref.read(budgetApiServiceProvider).addExpense(
-            widget.tripId,
-            category: draft.category,
-            label: draft.label,
-            amount: draft.amount,
-            sourceKind: source.kind,
-            sourceId: source.id,
-          );
-      _invalidateBudget();
-      if (!mounted) return;
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content:
-            Text(l10n.budgetPromptAdded(formatMoney(draft.amount, currency))),
-        action: SnackBarAction(
-          label: l10n.tripUndo,
-          onPressed: () async {
-            try {
-              await ref
-                  .read(budgetApiServiceProvider)
-                  .deleteExpense(widget.tripId, added.id);
-              _invalidateBudget();
-            } catch (e) {
-              _showSnack(l10n.tripUndoFailed(friendlyError(l10n, e)));
-            }
-          },
-        ),
-      ));
-    } catch (e) {
-      if (!mounted) return;
-      if (e is ApiException && e.statusCode == 422) {
-        _showSnack(l10n.budgetPromptLimitReached);
-      } else {
-        _showSnack(friendlyError(l10n, e));
-      }
-    }
-  }
-
-  /// Unbook cleanup: delete the auto (system-managed) expense linked to any
-  /// of [ids]. Best-effort — never rolls back a successful unbook; a
-  /// taken-over (auto=false) expense is the traveler's and stays
-  /// (migration 00061's contract).
-  Future<void> _removeLinkedAutoExpense(List<String> ids) async {
-    try {
-      final expenses = await ref.read(expensesProvider(widget.tripId).future);
-      var removed = false;
-      for (final e in expenses) {
-        if (e.auto && e.sourceId != null && ids.contains(e.sourceId)) {
-          await ref
-              .read(budgetApiServiceProvider)
-              .deleteExpense(widget.tripId, e.id);
-          removed = true;
-        }
-      }
-      if (removed) _invalidateBudget();
-    } catch (_) {}
-  }
-
-  void _invalidateBudget() {
-    ref.invalidate(budgetProvider(widget.tripId));
-    ref.invalidate(expensesProvider(widget.tripId));
-  }
-
-  /// Persists a drag-reorder of the Bookings section's residual "Other" cards.
-  /// Optimistic: residual display order derives from _bookingTodos list order
-  /// (see _groupedBookings), so move the residual entries to the tail in their
-  /// new order — grouped todos are slot-matched by key, so nothing else shifts.
-  Future<void> _reorderResidual(
-      List<BookingTodo> residual, int oldIndex, int newIndex) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    if (newIndex > oldIndex) newIndex--;
-    if (newIndex == oldIndex) return;
-    final newOrder = List.of(residual);
-    newOrder.insert(newIndex, newOrder.removeAt(oldIndex));
-    final residualIds = {for (final t in residual) t.id};
-    final prev = _bookingTodos;
-    setState(() {
-      _bookingTodos = [
-        for (final t in _bookingTodos)
-          if (!residualIds.contains(t.id)) t,
-        ...newOrder,
-      ];
-    });
-    try {
-      await ref
-          .read(bookingTodosApiServiceProvider)
-          .reorderTodos(widget.tripId, [for (final t in newOrder) t.id]);
-    } catch (e) {
-      if (mounted) setState(() => _bookingTodos = prev);
-      _showSnack(l10n.tripReorderFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  /// The residual booking-todo cards (the Bookings section's "Other"
-  /// sub-group): todos that didn't match a city group. City-matched todos
-  /// render embedded inside their city groups instead. Only called with a
-  /// non-empty [residual] — the section hides the sub-group otherwise.
-  Widget _residualBookingsList(List<BookingTodo> residual, ThemeData theme) {
-    final l10n = context.l10n;
-    return ReorderableListView.builder(
-      shrinkWrap: true,
-      physics: const NeverScrollableScrollPhysics(),
-      padding: EdgeInsets.zero,
-      buildDefaultDragHandles: false,
-      itemCount: residual.length,
-      onReorder: (oldIndex, newIndex) =>
-          _reorderResidual(residual, oldIndex, newIndex),
-      itemBuilder: (context, i) {
-        final todo = residual[i];
-        // Drag stays on all widths here: for residual custom bookings the
-        // handle is the ONLY reorder mechanism (no kebab move entries).
-        final canDrag = !_readOnly && !_isOffline && residual.length > 1;
-        return Padding(
-          key: ValueKey(todo.id),
-          padding: const EdgeInsets.symmetric(vertical: 4),
-          child: BookingTodoCard(
-            todo: todo,
-            onBookedChanged: (v) => _setRowBooked(v, todo: todo),
-            onOpen: _openCallbackFor(todo),
-            openLabelOverride: _flightLegs.containsKey(todo.todoKey)
-                ? l10n.tripFindFlights
-                : null,
-            onEdit: todo.auto ? null : () => _editTodo(todo),
-            onDelete: todo.auto ? null : () => _deleteTodo(todo),
-            dragHandle: canDrag
-                ? ReorderableDragStartListener(
-                    index: i,
-                    child: Icon(
-                      Icons.drag_indicator,
-                      color: theme.colorScheme.onSurfaceVariant,
-                    ),
-                  )
-                : null,
-          ),
-        );
-      },
-    );
-  }
-
-  /// "Add details…" on an inline booking row: promotes the todo to a
-  /// confirmed accommodation/segment via the existing add-sheets, prefilled
-  /// from the todo. Confirmed records are what viewers see and what calendar
-  /// export, the Tonight caption, and map stay pins read — this is the
-  /// one-tap replacement for the retired Suggested-draft "Keep" flow. The
-  /// next drafts sync sees the leg covered by a confirmed row and prunes its
-  /// shadow draft server-side. Deliberately does NOT mark the todo booked —
-  /// booking state stays a separate, explicit action.
-  Future<void> _addDetailsFromTodo(BookingTodo todo) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    if (todo.kind == 'stay') {
-      final body = await showModalBottomSheet<Map<String, dynamic>>(
-        context: context,
-        isScrollControlled: true,
-        builder: (_) => AddStaySheet(
-          initialName: todo.title,
-          initialCheckIn: todo.departDate,
-          initialCheckOut: todo.returnDate,
-        ),
-      );
-      if (body == null) return;
-      try {
-        await ref
-            .read(accommodationsApiServiceProvider)
-            .add(widget.tripId, body);
-        await _load();
-      } catch (e) {
-        _showSnack(l10n.tripAddStayFailed(friendlyError(l10n, e)));
-      }
-      return;
-    }
-    // Transport: the todo model carries no origin/destination fields, but
-    // _deriveTodos always titles a leg 'A → B' — split it back apart. Mode
-    // prefers the row's per-leg override, else follows the provider the leg
-    // was derived with.
-    final parts = todo.title.split(' → ');
-    final trip = _trip;
-    // Set by the "Change airport" link, which closes this sheet; the airports
-    // sheet opens after the await rather than from the callback, so the two
-    // modals never overlap.
-    var changeAirport = false;
-    // Prefill what the app actually KNOWS, not the wire's best-effort search
-    // seed. On a leg with no recorded flight the departure opens BLANK rather
-    // than pre-filled with the arrival day, so the sheet never asks the
-    // traveler to confirm a departure date the app invented. A row with no
-    // entry here (a custom or agent-added one) keeps the stored value.
-    final known = _legDates[todo.todoKey];
-    final body = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (sheetContext) => AddSegmentSheet(
-        initialOrigin: parts.isNotEmpty ? parts.first : null,
-        initialDestination: parts.length > 1 ? parts[1] : null,
-        initialMode: todo.mode ??
-            switch (todo.provider) {
-              'ferry' => 'ferry',
-              'rome2rio' => trip == null ? null : _groundModeOf(trip),
-              _ => 'flight',
-            },
-        initialDepartDate: known == null
-            ? todo.departDate
-            : (known.depart == null ? null : _fmt(known.depart!)),
-        initialArriveDate: known?.arrive == null ? null : _fmt(known!.arrive!),
-        // A derived leg's endpoints are the trip's (its airports) or the
-        // itinerary's (its cities) — this form's job is the booking detail.
-        // Typing over them here used to post a second row that contradicted
-        // the one above it.
-        endpointsLocked: todo.auto,
-        onChangeAirport: todo.auto && todo.isHomeLeg && !_readOnly
-            ? () {
-                changeAirport = true;
-                Navigator.of(sheetContext).pop();
-              }
-            : null,
-      ),
-    );
-    if (changeAirport) {
-      await _openTripAirports();
-      return;
-    }
-    if (body == null) return;
-    try {
-      await ref
-          .read(transportApiServiceProvider)
-          .addSegment(widget.tripId, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripAddTransportFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  /// Opens the trip's own departure/return airports (specs/trip-endpoint-
-  /// airports) and saves what comes back. The server renames the two derived
-  /// legs in the same transaction — in place, so their booked state, per-leg
-  /// mode and any linked expense survive — which is why this reloads the trip
-  /// rather than patching a title locally.
-  ///
-  /// This exists because the page had no such control: the only affordance on
-  /// a derived "EWR → Amsterdam" row was "Add details…", which posts a segment,
-  /// so correcting the airport there produced a second, contradicting row.
-  Future<void> _openTripAirports() async {
-    if (_guardOffline()) return;
-    final trip = _trip;
-    if (trip == null) return;
-    final l10n = context.l10n;
-    // What the legs read today when the trip states no airport of its own —
-    // shown as context, never seeded into the fields, so opening the sheet and
-    // pressing Save can't quietly promote a fallback into a fixed choice.
-    final fallback = (trip.origin?.trim().isNotEmpty ?? false)
-        ? trip.origin!.trim()
-        : _homeAirport;
-    final choice = await showModalBottomSheet<TripAirportsChoice>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => TripAirportsSheet(
-        originAirport: trip.originAirport,
-        returnAirport: trip.returnAirport,
-        fallbackLabel: fallback,
-      ),
-    );
-    if (choice == null) return;
-    try {
-      final result = await ref.read(tripsApiServiceProvider).updateTripEndpoints(
-            widget.tripId,
-            originAirport: choice.originAirport,
-            returnAirport: choice.returnAirport,
-          );
-      await _load();
-      if (!mounted) return;
-      _showSnack(l10n.tripAirportsSaved(result.legsRenamed.length));
-    } on TripEndpointsException catch (e) {
-      // The 422s are written for travelers and name what went wrong — an
-      // airport that matches nothing, or a trip that plainly travels by car.
-      // Flattening them into "something went wrong" throws away the answer.
-      _showSnack(e.statusCode == 422 && e.message.isNotEmpty
-          ? e.message
-          : l10n.tripAirportsFailed(friendlyError(l10n, e)));
-    } catch (e) {
-      _showSnack(l10n.tripAirportsFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  /// The per-leg mode writer: PATCHes the transport row's override (the
-  /// server stores it and rebuilds the row's provider + search link to
-  /// match), then re-derives the checklist so [_flightLegs]/[_ferryLegs] and
-  /// the row's open action agree with the new mode. Origin/destination come
-  /// from the derived 'A → B' title, exactly like [_addDetailsFromTodo].
-  Future<void> _setRowMode(BookingTodo todo, String mode) async {
-    if (_guardOffline()) return;
-    if (mode == todo.mode) return;
-    final l10n = context.l10n;
-    final parts = todo.title.split(' → ');
-    if (parts.length < 2) return;
-    try {
-      final updated = await ref.read(bookingTodosApiServiceProvider).setMode(
-            widget.tripId,
-            todo.id,
-            mode: mode,
-            origin: parts.first,
-            destination: parts[1],
-            departDate: todo.departDate,
-          );
-      if (!mounted) return;
-      setState(() => _bookingTodos = [
-            for (final t in _bookingTodos) t.id == updated.id ? updated : t
-          ]);
-      // A walk-derived Next Step reads this row's mode for its copy and its
-      // action label (specs/next-step-cta), and the sync below only
-      // invalidates when the DERIVED set changed — which a mode-only edit
-      // does not. Re-read the review explicitly so card and row agree.
-      _invalidateReview();
-      final trip = _trip;
-      if (trip != null) await _syncBookingTodos(trip);
-    } catch (e) {
-      _showSnack(l10n.tripUpdateFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _deleteTodo(BookingTodo todo) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    try {
-      await ref
-          .read(bookingTodosApiServiceProvider)
-          .delete(widget.tripId, todo.id);
-      if (mounted) {
-        setState(() => _bookingTodos =
-            _bookingTodos.where((t) => t.id != todo.id).toList());
-      }
-    } catch (e) {
-      _showSnack(l10n.tripDeleteFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _addBooking() async {
-    if (_guardOffline()) return;
-    final added = await showDialog<bool>(
-      context: context,
-      builder: (_) => _AddBookingTodoDialog(
-        tripId: widget.tripId,
-        groundMode: _trip == null ? null : _groundModeOf(_trip!),
-      ),
-    );
-    if (added == true) await _load();
-  }
-
-  Future<void> _editTodo(BookingTodo todo) async {
-    if (_guardOffline()) return;
-    final changed = await showDialog<bool>(
-      context: context,
-      builder: (_) =>
-          _AddBookingTodoDialog(tripId: widget.tripId, existing: todo),
-    );
-    if (changed == true) await _load();
-  }
-
   /// [day] preselects the dialog's Day dropdown (e.g. from the map's
   /// empty-day CTA, where the user is already looking at a specific day).
   Future<void> _addPlace({int? day}) async {
@@ -2030,70 +1524,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   }
 
   /// The trip page's visible proof that a conversation is waiting
-  /// (specs/trip-refine-memory). Without it, resuming would be an invisible
-  /// behavior of a button the traveler has to guess at — which is how the
-  /// conversation got lost in the first place.
-  ///
-  /// Hidden offline: reopening needs a fetch, and a dead row is worse than
-  /// none. Hidden while the panel is open — the conversation is right there.
-  Widget _continueChatRow(Trip trip, ThemeData theme, AppLocalizations l10n) {
-    final chat = trip.refineChat;
-    if (chat == null || _panelOpen || !trip.canEdit || _isOffline) {
-      return const SizedBox.shrink();
-    }
-    final updated = DateTime.tryParse(chat.updatedAt);
-    return Padding(
-      padding: const EdgeInsets.only(top: AppSpacing.md),
-      child: Card(
-        margin: EdgeInsets.zero,
-        child: ListTile(
-          leading: const Icon(Icons.forum_outlined),
-          title: Text(l10n.tripContinueChat,
-              style: theme.textTheme.titleSmall
-                  ?.copyWith(fontWeight: FontWeight.w600)),
-          subtitle: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              if (chat.preview.isNotEmpty)
-                Text(chat.preview, maxLines: 2, overflow: TextOverflow.ellipsis),
-              Padding(
-                padding: const EdgeInsets.only(top: AppSpacing.xs),
-                child: Text(
-                  // relativeTime already exists (offline_banner.dart) — reuse
-                  // it rather than growing a second "how long ago" rule.
-                  updated == null
-                      ? l10n.tripContinueChatMeta(chat.messageCount, '')
-                      : l10n.tripContinueChatMeta(
-                          chat.messageCount, relativeTime(l10n, updated)),
-                  style: theme.textTheme.bodySmall
-                      ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-                ),
-              ),
-            ],
-          ),
-          // A menu, not a bare icon: this sits a thumb's width from the row's
-          // own onTap, which OPENS the conversation, and discarding one must
-          // never be the near miss of resuming it. It is also the only way to
-          // be rid of a saved chat without first opening it and waiting out a
-          // full restore just to throw the transcript away.
-          trailing: PopupMenuButton<String>(
-            onSelected: (_) => _newChat(trip),
-            itemBuilder: (context) => [
-              PopupMenuItem(
-                value: 'clear',
-                child: ListTile(
-                  leading: const Icon(Icons.delete_outline),
-                  title: Text(l10n.refineClearChat),
-                  contentPadding: EdgeInsets.zero,
-                ),
-              ),
-            ],
-          ),
-          onTap: () => _openChat(trip),
-        ),
-      ),
-    );
-  }
 
   /// Makes sure the trip's saved conversation is loaded before anything is
   /// sent into it (specs/trip-refine-memory). Returns false when the caller
@@ -2534,114 +1964,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     }
   }
 
-  Future<void> _addStay() async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final body = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => const AddStaySheet(),
-    );
-    if (body == null) return;
-    try {
-      await ref
-          .read(accommodationsApiServiceProvider)
-          .add(widget.tripId, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripAddStayFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _deleteStay(Accommodation a) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    try {
-      await ref
-          .read(accommodationsApiServiceProvider)
-          .delete(widget.tripId, a.id);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripRemoveStayFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  /// Opens the stay sheet prefilled; a save PATCHes the row, which also
-  /// confirms it if it was a Suggested draft.
-  Future<void> _editStay(Accommodation a) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final body = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => AddStaySheet(initial: a),
-    );
-    if (body == null) return;
-    try {
-      await ref
-          .read(accommodationsApiServiceProvider)
-          .update(widget.tripId, a.id, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripUpdateStayFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _addSegment() async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    // A stated travel mode prefills the form ('mixed' doesn't pick a side).
-    final tm = _trip?.travelMode;
-    final body = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => AddSegmentSheet(initialMode: tm == 'mixed' ? null : tm),
-    );
-    if (body == null) return;
-    try {
-      await ref
-          .read(transportApiServiceProvider)
-          .addSegment(widget.tripId, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripAddTransportFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _deleteSegment(TripSegment s) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    try {
-      await ref
-          .read(transportApiServiceProvider)
-          .deleteSegment(widget.tripId, s.id);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripRemoveTransportFailed(friendlyError(l10n, e)));
-    }
-  }
-
-  /// Opens the transport sheet prefilled; a save PATCHes the row, which also
-  /// confirms it if it was a Suggested draft.
-  Future<void> _editSegment(TripSegment s) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final body = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => AddSegmentSheet(initial: s),
-    );
-    if (body == null) return;
-    try {
-      await ref
-          .read(transportApiServiceProvider)
-          .updateSegment(widget.tripId, s.id, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripUpdateTransportFailed(friendlyError(l10n, e)));
-    }
-  }
-
   /// Where the app is mounted on its host: '/' in dev, '/app/' in the
   /// Anchors the app-bar share menu so the iPad share popover has a rect to
   /// point at (share_plus requires sharePositionOrigin there).
@@ -2840,17 +2162,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   String _fmt(DateTime d) =>
       '${d.year}-${d.month.toString().padLeft(2, '0')}-${d.day.toString().padLeft(2, '0')}';
 
-  Widget _itemLeading(String? category, int position) {
-    switch (category) {
-      case 'restaurant':
-        return const CircleAvatar(child: Icon(Icons.restaurant, size: 18));
-      case 'attraction':
-        return const CircleAvatar(child: Icon(Icons.attractions, size: 18));
-      default:
-        return CircleAvatar(child: Text('${position + 1}'));
-    }
-  }
-
   /// A stored title is "long" when it's really the AI summary (multi-line or
   /// lengthy); such trips get a computed display title instead. The rule is
   /// shared (utils/trip_format.dart) so the trips list agrees with this header.
@@ -2896,185 +2207,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     return label;
   }
 
-  /// The group an item belongs to: its day-trip hub city when set, else its own
-  /// city (the shared rule in utils/trip_legs.dart).
-  String? _hubOf(ItineraryItem item) => hubOf(item);
-
-  /// The city an item belongs to: the AI-assigned [ItineraryItem.city] when set,
-  /// otherwise a best-effort parse of the formatted address (shared rule in
-  /// utils/trip_legs.dart).
-  String? _cityOf(ItineraryItem item) => cityOf(item);
-
-  /// Renders a hub group's items as slivers, split into "Day N" sub-sections
-  /// when items carry day numbers (day-trip batching applied within each day).
-  /// Each day is a [MultiSliver] whose header pins below the city header while
-  /// the day's items scroll past, then is pushed off by the next day. Legacy
-  /// items with no day fall back to flat day-trip batching with no day headers.
-  ///
-  /// [cityKey] is the display city (weather lookup, refine target); [groupKey]
-  /// identifies this run of it and prefixes the day keys, so a revisited city's
-  /// two runs never share a day header (see [_buildGroups]).
-  List<Widget> _buildGroupItemSlivers(String cityKey, String groupKey,
-      List<ItineraryItem> items, ThemeData theme, DateTime? tripStart,
-      {required bool showTonight,
-      ({DateTime? start, DateTime? end})? range,
-      List<int> emptyDays = const []}) {
-    // City-filler suppression happens here — NOT in the derivation's
-    // filtered list — so an all-filler city keeps its group (city header +
-    // embedded booking rows; the slot<->group mapping indexes over the full
-    // itinerary).
-    items = items.where((it) => !isCityFiller(it)).toList();
-    if (!items.any((it) => it.day != null)) {
-      // A dayless leg renders flat, with no day headers — so it has no gaps to
-      // point at. The derivation mirrors this branch and hands back an empty
-      // [emptyDays] here.
-      return _dayTripSectionSlivers(items, theme);
-    }
-    // Per-day weather (specs/weather-in-itinerary): one report per city group
-    // for its date window — one API call per city per trip view, cached by the
-    // family provider. Best-effort: valueOrNull stays null until it resolves
-    // (and forever on failure), so nothing renders and the pinned-scroll math
-    // is untouched. 'Other places' has no real city to geocode. The watch
-    // itself lives in each day's chip Consumer (see _weatherChipSliver) so a
-    // report resolving repaints those chips, never the whole screen.
-    final weatherQuery = (cityKey != _kOtherPlaces &&
-            range?.start != null &&
-            range?.end != null)
-        ? WeatherQuery(
-            city: cityKey,
-            startDate: _fmt(range!.start!),
-            endDate: _fmt(range.end!),
-          )
-        : null;
-    // Today mode: the header for today's trip day (if any) gets a visible
-    // highlight; undated/past/future trips resolve to null and render as-is.
-    final todayDay =
-        tripDayOn(_trip?.startDate, _trip?.endDate, DateTime.now());
-    final slivers = <Widget>[];
-    // Running month for the narrow day labels: the month is spelled out when
-    // it is the first dated day of this group and whenever it CHANGES, and
-    // dropped in between. So a city reads "Sat, Aug 29 · Sun 30 · Mon 31 ·
-    // Tue, Sep 1 · Wed 2" — the way a person writes a list of dates. Dropping
-    // it unconditionally would leave "Sun 31 / Tue 2" genuinely ambiguous
-    // across a month rollover, which is exactly where a traveler is checking.
-    int? lastMonth;
-
-    // Open days interleave with the planned ones in day order. They are
-    // disjoint from the item days by construction (both come from one
-    // planned-day set in TripDerivation.compute), so this merge can never
-    // double-render a day; anything left over flushes after the loop.
-    var nextEmpty = 0;
-    void flushEmptiesBefore(int? day) {
-      while (nextEmpty < emptyDays.length &&
-          (day == null || emptyDays[nextEmpty] < day)) {
-        final d = emptyDays[nextEmpty++];
-        slivers.add(SliverToBoxAdapter(
-          child: _emptyDayRow(d, '$groupKey#$d', tripStart, theme,
-              isToday: d == todayDay,
-              onPlan: _planDaysAction(groupKey, cityKey, [d])),
-        ));
-      }
-    }
-
-    var i = 0;
-    while (i < items.length) {
-      final day = items[i].day;
-      final run = <ItineraryItem>[];
-      while (i < items.length && items[i].day == day) {
-        run.add(items[i]);
-        i++;
-      }
-      if (day != null) {
-        flushEmptiesBefore(day);
-        final dayKey = '$groupKey#$day';
-        final collapsed = _collapsedDays.contains(dayKey);
-        final month = tripStart?.add(Duration(days: day - 1)).month;
-        final showMonth = month == null || month != lastMonth;
-        lastMonth = month;
-        final header = _daySubHeader(
-            day, tripStart, theme, collapsed, _runTravelMin(run), () {
-          setState(() {
-            if (collapsed) {
-              _collapsedDays.remove(dayKey);
-            } else {
-              _collapsedDays.add(dayKey);
-            }
-          });
-        },
-            // Refine needs the network; owners and editor co-planners both
-            // get the per-day refine icon (viewers don't).
-            (!_isOffline && (_trip?.canEdit ?? true))
-                ? () {
-                    final trip = _trip;
-                    if (trip == null) return;
-                    // 'Other places' is a fallback label, not a real hub —
-                    // omit the city qualifier so the server matches on the
-                    // day alone.
-                    _openRefine(
-                        trip,
-                        RefineTarget.day(day,
-                            city: cityKey == _kOtherPlaces ? null : cityKey));
-                  }
-                : null,
-            headerKey: _dayHeaderKeys.putIfAbsent(dayKey, GlobalKey.new),
-            isToday: day == todayDay,
-            showMonth: showMonth);
-        // Tonight caption (specs/happening-now): a non-pinned content row —
-        // it scrolls and collapses with the section, never joining the
-        // pinned chrome. [showTonight] is true for at most one group, so
-        // repeated day numbers across city groups can't duplicate it.
-        final trip = _trip;
-        final tonight = (showTonight && day == todayDay && trip != null)
-            ? _tonightCaption(theme, _derive(trip).staysOnNight(day))
-            : null;
-        // Weather chip for this day, if the report covers its date. Historical
-        // reports carry last year's month-day, so we match on month-day. The
-        // sliver exists whenever a lookup is possible and renders zero-size
-        // until (unless) the report resolves and covers the day — a Consumer,
-        // so the resolution repaints only this chip.
-        Widget? weatherChip;
-        if (weatherQuery != null && tripStart != null) {
-          final md = _fmt(tripStart.add(Duration(days: day - 1))).substring(5);
-          weatherChip = Consumer(
-            builder: (context, ref, _) {
-              final report = ref.watch(weatherByCityProvider(weatherQuery)
-                  .select((a) => a.valueOrNull));
-              final wd = report?.dayFor(md);
-              if (wd == null) return const SizedBox.shrink();
-              return _weatherChip(theme, wd, report!.isHistorical);
-            },
-          );
-        }
-        // A collapsed day pins nothing: pinning exists so a header stays
-        // visible while its content scrolls beneath it, and a zero-body
-        // pinned group's paintOrigin (= viewport overlap) poisons
-        // MultiSliver's minPaintOrigin — the header squishes, then
-        // vanishes, as the pinned chrome scrolls in. Same hazard as the
-        // collapsed city groups in build.
-        if (collapsed) {
-          slivers.add(SliverToBoxAdapter(child: header));
-        } else {
-          slivers.add(MultiSliver(
-            pushPinnedChildren: true,
-            children: [
-              SliverPinnedHeader(child: header),
-              if (weatherChip != null) _boxSliver([weatherChip]),
-              if (tonight != null) _boxSliver([tonight]),
-              ..._dayTripSectionSlivers(run, theme),
-            ],
-          ));
-        }
-      } else {
-        slivers.addAll(_dayTripSectionSlivers(run, theme));
-      }
-    }
-    // Open days after the leg's last planned one — the common case on a spine,
-    // where the middle of the stay is empty and only the arrival and move-on
-    // days carry anything.
-    flushEmptiesBefore(null);
-    return slivers;
-  }
-
   /// Wraps a run of box widgets as a single sliver for use inside MultiSliver.
   Widget _boxSliver(List<Widget> children) => SliverToBoxAdapter(
         child: Column(
@@ -3082,938 +2214,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
           children: children,
         ),
       );
-
-  /// Shared width for every city-header date chip this build: max over groups
-  /// of icon + range + gap + nights, measured with the exact strings and the
-  /// exact style the chips render, capped at [_chipMaxWidth]. One width for
-  /// all rows is what makes the chips align as columns. Compute it as a
-  /// build-local alongside the groups list — never cache it in state: pinned
-  /// headers rebuild from the same pass as scrolling rows, and a stale cached
-  /// width would misalign them against fresh ones.
-  /// The one chip text style, shared by the chip Texts and the measurement
-  /// so the two cannot drift.
-  TextStyle? _chipTextStyle(ThemeData theme) =>
-      theme.textTheme.labelMedium?.copyWith(color: theme.colorScheme.primary);
-
-  double _dateChipWidth(List<CityGroup> groups, ThemeData theme) {
-    var style = _chipTextStyle(theme);
-    // Text applies the boldText accessibility flag internally; TextPainter
-    // does not — merge it here or measurement undershoots the rendered width.
-    if (MediaQuery.boldTextOf(context)) {
-      style = style?.copyWith(fontWeight: FontWeight.bold);
-    }
-    final tp = TextPainter(
-      textDirection: Directionality.of(context),
-      // The TextScaler object, not a factor: Android 14+ scaling is nonlinear.
-      textScaler: MediaQuery.textScalerOf(context),
-      maxLines: 1,
-    );
-    double measure(String s) {
-      tp.text = TextSpan(text: s, style: style);
-      tp.layout();
-      // Whole px: a float-exact fit on the widest row would self-ellipsize.
-      return tp.width.ceilToDouble();
-    }
-
-    var w = 0.0;
-    for (final g in groups) {
-      final chip = g.dateRange;
-      if (chip == null) continue;
-      var cw = _chipIconSpan + measure(chip.range);
-      if (chip.nights != null) cw += _chipInnerGap + measure(chip.nights!);
-      if (cw > w) w = cw;
-    }
-    tp.dispose();
-    return w > _chipMaxWidth ? _chipMaxWidth : w;
-  }
-
-  /// City group header: name, date range, refine + collapse controls. Pinned
-  /// at the top of the scroll area while its group scrolls past; the opaque
-  /// Material keeps items from showing through while pinned. [cityCollapsed]
-  /// comes from the caller's render branch — the one _collapsedGroups read.
-  Widget _cityHeader(Trip trip, CityGroup group, ThemeData theme,
-      double dateChipWidth,
-      {required bool cityCollapsed}) {
-    final l10n = context.l10n;
-    final chipStyle = _chipTextStyle(theme);
-    // Narrow drops the dates OUT of the header row and stacks them under the
-    // name (see [_cityHeaderMetaLine]). The row's rigid chip is measured for
-    // the widest group and can claim up to [_chipMaxWidth] of a ~300px phone
-    // row, and since the label's Expanded is the only flex child it absorbs
-    // the whole deficit — which is how "Prague" rendered as "Pra…". Columns
-    // are what the chip buys, and a phone has no room to spend on them.
-    final stackDates = _narrow && group.dateRange != null;
-    final header = HoverReveal(
-      builder: (context, revealed) => Material(
-        color: theme.scaffoldBackgroundColor,
-        child: InkWell(
-          onTap: () {
-            // A pure list toggle: expansion is decoupled from the map, so a
-            // header tap never writes focus, never moves the camera, and
-            // never touches any other group. No scroll either — the user is
-            // already at the header.
-            setState(() {
-              cityCollapsed
-                  ? _collapsedGroups.remove(group.key)
-                  : _collapsedGroups.add(group.key);
-            });
-          },
-          child: Padding(
-            // Narrow buys back some of the second line's height at the top,
-            // where the gap between a city and the day header under it was
-            // the loosest thing on the phone.
-            padding: _narrow
-                ? const EdgeInsets.fromLTRB(16, 12, 16, 4)
-                : const EdgeInsets.fromLTRB(16, 16, 16, 4),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Icon(Icons.location_on,
-                        size: 18, color: theme.colorScheme.primary),
-                    const SizedBox(width: 6),
-                    // The qualifier rides INSIDE this Expanded, never as a
-                    // second flex child: a sibling Flexible would split the
-                    // free space with the label and drag the chip+chevron
-                    // cluster off the right edge (see the date chip's note
-                    // below, PR #307). It stacks BELOW the label rather than
-                    // beside it, so the label's Text keeps the outer Row as its
-                    // nearest Row ancestor — the header tests navigate that
-                    // ancestry to pair a city with its date chip.
-                    Expanded(
-                      child: _cityHeaderLabel(l10n, group, theme,
-                          metaLine: stackDates
-                              ? _cityHeaderMetaLine(group, chipStyle)
-                              : null),
-                    ),
-                    if (group.dateRange != null && !stackDates)
-                      // A rigid SizedBox at the per-build shared width
-                      // ([_dateChipWidth]), so calendar icons, range starts,
-                      // and nights suffixes form columns across rows.
-                      // Deliberately NOT Flexible: a second flex child would
-                      // split the free space with the label's Expanded and
-                      // drag the chip+chevron cluster off the right edge
-                      // (PR #307) — the inner Expanded on the range lives
-                      // inside the chip's own Row, invisible to the outer
-                      // flex accounting, and absorbs any deficit by
-                      // ellipsizing when the shared width hits the
-                      // pathological [_chipMaxWidth] cap; the nights suffix
-                      // keeps its intrinsic width flush right.
-                      MergeSemantics(
-                        // One utterance per chip ("Aug 24 – Aug 27 · 3
-                        // nights"), matching the pre-split reading order.
-                        child: SizedBox(
-                          width: dateChipWidth,
-                          child: Row(
-                            children: [
-                              Icon(Icons.event,
-                                  size: _chipIconSize,
-                                  color: theme.colorScheme.primary),
-                              const SizedBox(width: _chipIconGap),
-                              Expanded(
-                                child: Text(
-                                  group.dateRange!.range,
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                  style: chipStyle,
-                                ),
-                              ),
-                              if (group.dateRange!.nights != null) ...[
-                                const SizedBox(width: _chipInnerGap),
-                                ConstrainedBox(
-                                  // Guard: at extreme accessibility scale a
-                                  // nights label alone could exceed the
-                                  // capped chip — bound it so it ellipsizes
-                                  // instead of overflowing the inner Row.
-                                  constraints: const BoxConstraints(
-                                      maxWidth: _chipMaxWidth -
-                                          _chipIconSpan -
-                                          _chipInnerGap),
-                                  child: Text(
-                                    group.dateRange!.nights!,
-                                    maxLines: 1,
-                                    overflow: TextOverflow.ellipsis,
-                                    style: chipStyle,
-                                  ),
-                                ),
-                              ],
-                            ],
-                          ),
-                        ),
-                      ),
-                    // 'Other places' has no hub the section tool can target;
-                    // refine also needs the network. Hover-revealed on
-                    // pointer devices (always visible on touch) to keep the
-                    // resting row quiet. canEdit/offline are screen-wide —
-                    // when they drop the button they drop it on EVERY row, so
-                    // the chip columns stay aligned. 'Other places' is the
-                    // one per-row variance: it keeps an invisible,
-                    // width-identical placeholder (same IconButton config, so
-                    // density changes can't drift it) or its chip cluster
-                    // would sit a button-slot right of the other rows —
-                    // EXCEPT on narrow, where the dates have left the row and
-                    // there is no cluster left to align. Holding ~40px of a
-                    // phone row for a button nobody can see is the density
-                    // this pass exists to remove.
-                    if (trip.canEdit && !_isOffline)
-                      group.label != _kOtherPlaces
-                          ? HoverRevealed(
-                              revealed: revealed,
-                              child: IconButton(
-                                icon: const Icon(Icons.auto_awesome, size: 16),
-                                tooltip: l10n.tripRefineCity(group.label),
-                                visualDensity: VisualDensity.compact,
-                                color: theme.colorScheme.primary,
-                                onPressed: () => _openRefine(
-                                    trip, RefineTarget.city(group.label)),
-                              ),
-                            )
-                          : _narrow
-                              ? const SizedBox.shrink()
-                              : ExcludeSemantics(
-                                  child: IgnorePointer(
-                                    child: Opacity(
-                                      opacity: 0,
-                                      // No tooltip: IgnorePointer blocks
-                                      // hover, a ghost tooltip target would
-                                      // outlive it.
-                                      child: IconButton(
-                                        icon: const Icon(Icons.auto_awesome,
-                                            size: 16),
-                                        visualDensity: VisualDensity.compact,
-                                        onPressed: null,
-                                      ),
-                                    ),
-                                  ),
-                                ),
-                    const SizedBox(width: 4),
-                    Icon(
-                      cityCollapsed ? Icons.chevron_right : Icons.expand_more,
-                      size: 20,
-                      color: theme.colorScheme.primary,
-                    ),
-                  ],
-                ),
-                // Open days, promoted (specs/shape-before-schedule): a visible,
-                // labelled version of the same city refine the hover sparkle
-                // above performs — one action, two affordances. It is a second
-                // LINE and never a change to the Row, because the Row's widths
-                // are what align the date chips into columns across cities (see
-                // the ghost-placeholder note above); a button in there would
-                // shift every other city's chip.
-                //
-                // Renders on the collapsed header too, which is exactly where
-                // "2 days unplanned" earns its place.
-                if (_planDaysAction(group.key, group.label, group.emptyDays)
-                    case final plan?) ...[
-                  const SizedBox(height: 2),
-                  Padding(
-                    padding: const EdgeInsets.only(left: 24),
-                    child: Row(
-                      children: [
-                        Icon(Icons.edit_calendar_outlined,
-                            size: 14, color: theme.colorScheme.onSurfaceVariant),
-                        const SizedBox(width: 6),
-                        Flexible(
-                          child: Text(
-                            l10n.tripUnplannedDays(group.emptyDays.length),
-                            style: theme.textTheme.bodySmall?.copyWith(
-                                color: theme.colorScheme.onSurfaceVariant),
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                        // Flexible + ellipsis for the same reason the label
-                        // beside it is: Spanish at 1.3x on a 320px viewport
-                        // overflowed this Row while the button was its one
-                        // unflexible child.
-                        Flexible(
-                          child: TextButton(
-                            key: ValueKey('plan-days:${group.key}'),
-                            style: TextButton.styleFrom(
-                                visualDensity: VisualDensity.compact,
-                                padding:
-                                    const EdgeInsets.symmetric(horizontal: 8)),
-                            onPressed: plan,
-                            child: Text(
-                              l10n.tripPlanTheseDays,
-                              maxLines: 1,
-                              overflow: TextOverflow.ellipsis,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-                const SizedBox(height: 4),
-                const Divider(height: 1),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-    // Repaint-isolate the header: the open city's header rides pinned over
-    // the scrolling list (and collapsed ones scroll with it), and the hover
-    // setState + InkWell highlight must repaint this header's own layer, not
-    // re-record the viewport picture. Covers pinned and collapsed branches.
-    return RepaintBoundary(child: header);
-  }
-
-  /// Curated, locally-sourced recommendations for a city group — vetted picks
-  /// from real locals. Renders nothing when there is no coverage for the city
-  /// (empty list) or on error, so it never shows a broken/empty section.
-  Widget _localIntelSliver(String label, ThemeData theme) {
-    if (label == _kOtherPlaces) {
-      return const SliverToBoxAdapter(child: SizedBox.shrink());
-    }
-    return SliverToBoxAdapter(
-      child: Consumer(builder: (context, ref, _) {
-        final recs =
-            ref.watch(localRecsByCityProvider(label)).valueOrNull ?? [];
-        final guides =
-            ref.watch(localGuidesByCityProvider(label)).valueOrNull ?? [];
-        if (recs.isEmpty && guides.isEmpty) return const SizedBox.shrink();
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Padding(
-              // left: lines the icon up with the booking rows' kind icons,
-              // which sit at 12px (BookingTodoRow's own left padding).
-              padding: const EdgeInsets.only(
-                  left: AppSpacing.md, top: AppSpacing.sm, bottom: AppSpacing.xs),
-              child: Row(
-                children: [
-                  Icon(Icons.verified, size: 16, color: AppColors.toolLocal),
-                  const SizedBox(width: 6),
-                  Text(
-                    context.l10n.tripLocalIntel,
-                    style: theme.textTheme.labelLarge?.copyWith(
-                      fontWeight: FontWeight.w600,
-                      color: AppColors.toolLocal,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            for (final g in guides) _guideChip(g, theme),
-            for (final r in recs.take(6))
-              LocalRecCard(
-                rec: r,
-                onAddToTrip: () =>
-                    _addToTrip(AddToTripPayload.fromLocalRec(r)),
-              ),
-          ],
-        );
-      }),
-    );
-  }
-
-  /// A tappable "Local guide" row inside the Local intel section that opens the
-  /// full narrative guide (story + ordered pins + map).
-  Widget _guideChip(LocalGuide guide, ThemeData theme) {
-    final accent = AppColors.toolLocal;
-    return Card(
-      margin: const EdgeInsets.symmetric(vertical: AppSpacing.xs),
-      clipBehavior: Clip.antiAlias,
-      child: InkWell(
-        onTap: () => pushOnce(
-            context,
-            MaterialPageRoute(
-              builder: (_) => LocalGuideDetailScreen(guide: guide),
-            )),
-        child: Padding(
-          padding: const EdgeInsets.all(AppSpacing.md),
-          child: Row(
-            children: [
-              Container(
-                padding: const EdgeInsets.all(AppSpacing.sm),
-                decoration: BoxDecoration(
-                  color: accent.withValues(alpha: 0.12),
-                  borderRadius: BorderRadius.circular(AppRadius.sm),
-                ),
-                child: Icon(Icons.menu_book, size: 20, color: accent),
-              ),
-              const SizedBox(width: AppSpacing.md),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      context.l10n.tripLocalGuideTitle(guide.title),
-                      style: theme.textTheme.titleSmall
-                          ?.copyWith(fontWeight: FontWeight.w600),
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    if (guide.sourceName.isNotEmpty) ...[
-                      const SizedBox(height: 2),
-                      Text(
-                        context.l10n.tripGuideBy(guide.sourceName),
-                        style: theme.textTheme.bodySmall?.copyWith(
-                            color: accent, fontWeight: FontWeight.w600),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-              const SizedBox(width: AppSpacing.sm),
-              Icon(Icons.chevron_right,
-                  color: theme.colorScheme.onSurfaceVariant),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  /// Live local-events section for a city group, looked up for the group's
-  /// date window and rendered as the same horizontal poster rail the plan chat
-  /// uses for these events ([PlacePhotoStrip] + [PlaceCardData.event]) — one
-  /// card system, two surfaces. Returns an empty box (no sliver content) when
-  /// the group has no real city/dates to query. Wrapped in a [Consumer] so
-  /// only this section rebuilds as the async lookup resolves.
-  ///
-  /// [range] is the group's VISIBLE range — the window the city header chip
-  /// shows the traveler. It has to be: a section headed "Events while you're
-  /// here" that queried a different window than the one on screen is what made
-  /// a Sep 1–4 Berlin leg show five events, all on Sep 4.
-  Widget _eventsSliver(
-    Trip trip,
-    String label,
-    ({DateTime? start, DateTime? end})? range,
-    ThemeData theme,
-  ) {
-    if (label == _kOtherPlaces ||
-        range?.start == null ||
-        range?.end == null) {
-      return const SliverToBoxAdapter(child: SizedBox.shrink());
-    }
-    final query = EventsQuery(
-      city: label,
-      startDate: _fmt(range!.start!),
-      endDate: _fmt(range.end!),
-    );
-    return SliverToBoxAdapter(
-      child: Consumer(builder: (context, ref, _) {
-        final async = ref.watch(eventsByCityProvider(query));
-        return async.when(
-          loading: () => Padding(
-            padding: const EdgeInsets.symmetric(vertical: AppSpacing.sm),
-            child: Row(
-              children: [
-                const SizedBox(
-                  width: 14,
-                  height: 14,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                ),
-                const SizedBox(width: AppSpacing.sm),
-                // Flexible, or a long city name overflows phones for the
-                // few frames this spinner row is on screen.
-                Flexible(
-                  child: Text(context.l10n.tripFindingEvents(label),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                      style: theme.textTheme.bodySmall?.copyWith(
-                          color: theme.colorScheme.onSurfaceVariant)),
-                ),
-              ],
-            ),
-          ),
-          // On error (e.g. provider key not set), try the Greek source-links
-          // fallback rather than going silent.
-          error: (_, __) => _greekEventsFallback(query, theme),
-          data: (events) {
-            if (events.isEmpty) return _greekEventsFallback(query, theme);
-            // Day-spread, not the chronological head: on a busy first night a
-            // plain take(N) fills every slot from day one and the rest of the
-            // stay never appears (utils/event_picks.dart).
-            final picks = spreadEventsByDay(events, limit: kEventRailCards);
-            return Padding(
-              // left: lines the strip's header icon up with the booking rows'
-              // kind icons, which sit at 12px (BookingTodoRow's own left
-              // padding) — the leading grid asserted by
-              // trip_detail_booking_alignment_test.
-              padding: const EdgeInsets.only(left: AppSpacing.md),
-              child: PlacePhotoStrip(
-                icon: Icons.local_activity,
-                accent: AppColors.toolEvents,
-                // Counts everything found, not the cards shown — otherwise 8
-                // reads as "that's all there is". At the server's per-city cap
-                // the true total is unknown, so the count says "30+".
-                label: events.length >= kEventsServerCap
-                    ? context.l10n
-                        .tripEventsWhileHereCountCapped(events.length)
-                    : context.l10n.tripEventsWhileHereCount(events.length),
-                actionLabel: context.l10n.commonSeeAll,
-                onViewTrip: events.length > picks.length
-                    ? () => showCityEventsSheet(
-                          context,
-                          city: label,
-                          events: events,
-                          onAddToTrip: (e) =>
-                              _addToTrip(AddToTripPayload.fromEvent(e)),
-                        )
-                    : null,
-                cards: [
-                  for (final e in picks)
-                    PlacePhotoCard(
-                      data: PlaceCardData.event(e),
-                      onTap: e.url.isEmpty
-                          ? null
-                          : () => trackedLaunchUrl(context, e.url,
-                              provider: 'ticketmaster',
-                              surface: 'trip_event_card',
-                              tripId: trip.id),
-                      onAddToTrip: () =>
-                          _addToTrip(AddToTripPayload.fromEvent(e)),
-                    ),
-                ],
-              ),
-            );
-          },
-        );
-      }),
-    );
-  }
-
-  /// When the structured events lookup is empty/errored, show curated Greek
-  /// event-discovery links for Greek cities (empty for everywhere else, so this
-  /// renders nothing). Keeps the section useful where Ticketmaster has no data.
-  Widget _greekEventsFallback(EventsQuery query, ThemeData theme) {
-    return Consumer(builder: (context, ref, _) {
-      final links = ref.watch(greeceEventLinksProvider(query)).valueOrNull;
-      if (links == null || links.isEmpty) return const SizedBox.shrink();
-      return SourceLinksCard(
-        icon: Icons.local_activity,
-        accent: AppColors.toolEvents,
-        title: context.l10n.tripFindEventsIn(query.city),
-        links: links,
-      );
-    });
-  }
-
-  /// Greek islands/ports we offer ferry connectors between (mirrors the API's
-  /// isGreekLocation set; kept small and local since it only gates the UI hint).
-  static const _greekIslands = {
-    'athens', 'piraeus', 'santorini', 'thira', 'fira', 'oia', 'mykonos',
-    'naxos', 'paros', 'ios', 'milos', 'syros', 'tinos', 'folegandros',
-    'crete', 'heraklion', 'chania', 'rethymno', 'rhodes', 'kos', 'corfu',
-    'kefalonia', 'zakynthos', 'lefkada', 'skiathos', 'skopelos', 'samos',
-    'chios', 'lesbos', 'mytilene', 'karpathos', 'symi', 'hydra', 'spetses',
-    'aegina',
-  };
-
-  bool _isGreekIsland(String label) {
-    final n = label.toLowerCase().trim();
-    if (n.contains('greece')) return true;
-    if (_greekIslands.contains(n)) return true;
-    final comma = n.indexOf(',');
-    if (comma > 0 && _greekIslands.contains(n.substring(0, comma).trim())) {
-      return true;
-    }
-    return false;
-  }
-
-
-  /// The saved-details line under a slot's todo row (or standalone when the
-  /// slot has no todo — viewers get none from the server). Carries the
-  /// confirmed record's edit/delete/calendar affordances; the checkbox shows
-  /// only in detail-only mode, where there's no todo row above to drive it.
-  Widget _detailRowFor({Accommodation? stay, TripSegment? segment,
-      required bool detailOnly}) {
-    final editable = !_readOnly && !_isOffline;
-    if (stay != null) {
-      return BookingDetailRow.stay(
-        tripId: widget.tripId,
-        stay: stay,
-        onEdit: editable ? () => _editStay(stay) : null,
-        onDelete: editable ? () => _deleteStay(stay) : null,
-        showCheckbox: detailOnly,
-        onBookedChanged:
-            detailOnly && !_readOnly ? (v) => _setRowBooked(v, stay: stay) : null,
-        appleCalendarEnabled: !_readOnly && !_isOffline,
-      );
-    }
-    final s = segment!;
-    return BookingDetailRow.segment(
-      tripId: widget.tripId,
-      segment: s,
-      onEdit: editable ? () => _editSegment(s) : null,
-      onDelete: editable ? () => _deleteSegment(s) : null,
-      showCheckbox: detailOnly,
-      onBookedChanged:
-          detailOnly && !_readOnly ? (v) => _setRowBooked(v, segment: s) : null,
-      appleCalendarEnabled: !_readOnly && !_isOffline,
-    );
-  }
-
-  /// Compact booking rows for a city group's slot: arrival flight + stay when
-  /// [departureOnly] is false, the return-home flight when true. Each todo
-  /// row is followed by its matched confirmed record's details; a match
-  /// without a todo (viewers) renders as a standalone detail row.
-  /// [unbookedOnly] keeps only rows whose visible checkbox is unchecked — the
-  /// "Not booked yet" lens.
-  ///
-  /// The slot's entries and their checkbox state both come from the
-  /// derivation ([bookingSlotEntries] / [bookingEntryBooked]) rather than
-  /// being spelled out here, because the filter-strip counts iterate the same
-  /// two functions: a chip's count and the rows it reveals cannot disagree
-  /// about what a slot holds or about what "booked" means.
-  List<Widget> _bookingRowWidgets(
-    BookingSlot slot, {
-    required bool departureOnly,
-    bool unbookedOnly = false,
-  }) {
-    final l10n = context.l10n;
-    var entries = bookingSlotEntries(slot, departureOnly: departureOnly);
-    if (unbookedOnly) {
-      entries = entries.where((e) => !bookingEntryBooked(e)).toList();
-    }
-    return [
-      for (final e in entries) ...[
-        if (e.todo case final todo?)
-          BookingTodoRow(
-            todo: todo,
-            tripTravelMode: _trip?.travelMode,
-            compact: _narrow,
-            onBookedChanged: (v) => _setRowBooked(v,
-                todo: todo, stay: e.stay, segment: e.segment),
-            onOpen: _openCallbackFor(todo),
-            openLabelOverride: _ferryLegs.containsKey(todo.todoKey)
-                ? (_narrow ? l10n.tripFindFerriesShort : l10n.tripFindFerries)
-                : _flightLegs.containsKey(todo.todoKey)
-                    ? (_narrow
-                        ? l10n.tripFindFlightsShort
-                        : l10n.tripFindFlights)
-                    : null,
-            // No "Add details…" once a confirmed segment fills the slot —
-            // the same rule as the mode picker below: that row's truth is the
-            // segment, edited via its own sheet. Without this the sheet would
-            // open pre-filled with the segment's own dates and Save would
-            // create a SECOND segment for the same leg. (Locking the sheet's
-            // endpoints stops a segment that CONTRADICTS the row; this stops a
-            // duplicate of it — the two guards cover different halves.)
-            onAddDetails: (_readOnly ||
-                    _isOffline ||
-                    todo.kind == 'other' ||
-                    e.segment != null)
-                ? null
-                : () => _addDetailsFromTodo(todo),
-            // Only the two journey endpoints: those are the rows the trip's
-            // own airports title, so they are the only ones this moves. The
-            // role comes from the server, which stores it as identity —
-            // guessing it here would be wrong on a row it demoted.
-            onChangeAirport: (_readOnly || _isOffline || !todo.isHomeLeg)
-                ? null
-                : () => _openTripAirports(),
-            // No picker when a confirmed segment fills the slot — that row's
-            // mode truth is the segment, edited via its own sheet.
-            onModeChanged: (_readOnly ||
-                    _isOffline ||
-                    e.segment != null ||
-                    todo.kind != 'transport')
-                ? null
-                : (m) => _setRowMode(todo, m),
-          ),
-        if (e.stay != null || e.segment != null)
-          _detailRowFor(
-              stay: e.stay, segment: e.segment, detailOnly: e.todo == null),
-      ],
-    ];
-  }
-
-  /// Flat "left to book" rows for the 'unbooked' lens: every city slot's
-  /// unbooked rows in trip order, then the residual todos and detail-only
-  /// records that matched no city. Reuses the same row widgets (and the
-  /// _setRowBooked writer) as the inline city view, so checking a box here
-  /// behaves identically and the row leaves the lens on the rebuild.
-  ///
-  /// [destination] narrows to one leg label exactly as [_allBookingRows] does
-  /// — the destination strip renders in BOTH scopes, so a chosen city has to
-  /// mean the same thing in both. Same discipline too: this filters the OUTPUT
-  /// of the one full-label partition, never a re-partition on a label subset.
-  List<Widget> _unbookedRows(
-    GroupedBookings grouped,
-    List<String> labels, {
-    String? destination,
-  }) {
-    final l10n = context.l10n;
-    bool slotShown(int i) =>
-        destination == null ||
-        (i < labels.length && labels[i] == destination);
-    final showResiduals =
-        destination == null || destination == _kOtherPlaces;
-    return [
-      for (final (i, slot) in grouped.slots.indexed)
-        if (slotShown(i)) ...[
-          ..._bookingRowWidgets(slot, departureOnly: false, unbookedOnly: true),
-          if (i == grouped.slots.length - 1)
-            ..._bookingRowWidgets(slot,
-                departureOnly: true, unbookedOnly: true),
-        ],
-      if (showResiduals) ...[
-        for (final todo in grouped.residual.where((t) => !t.booked))
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 4),
-            child: BookingTodoCard(
-              todo: todo,
-              onBookedChanged: (v) => _setRowBooked(v, todo: todo),
-              onOpen: _openCallbackFor(todo),
-              openLabelOverride: _flightLegs.containsKey(todo.todoKey)
-                  ? l10n.tripFindFlights
-                  : null,
-              onEdit: todo.auto ? null : () => _editTodo(todo),
-              onDelete: todo.auto ? null : () => _deleteTodo(todo),
-            ),
-          ),
-        for (final a in grouped.residualStays.where((a) => !a.booked))
-          _detailRowFor(stay: a, segment: null, detailOnly: true),
-        for (final s in grouped.residualSegments.where((s) => !s.booked))
-          _detailRowFor(stay: null, segment: s, detailOnly: true),
-      ],
-    ];
-  }
-
-  /// The 'bookings' lens rows: every city slot's rows — booked and unbooked —
-  /// in trip order, then the residual todos and detail-only records that
-  /// matched no city. Reuses the same row widgets (and the _setRowBooked
-  /// writer) as the inline city view, so checking a box here behaves
-  /// identically; the row stays put, struck through.
-  ///
-  /// [destination] narrows to one leg label (null = all): slot i is included
-  /// when labels[i] matches; residuals only under All or the 'Other places'
-  /// chip (they matched no destination by definition). This filters the
-  /// OUTPUT of the one full-label groupedBookings partition — never re-run
-  /// the partition on a label subset: its claim-once matching is
-  /// order-dependent, so a subset call would assign rows differently than
-  /// the inline city view (docs/zen.md).
-  List<Widget> _allBookingRows(
-    GroupedBookings grouped,
-    List<String> labels, {
-    String? destination,
-  }) {
-    final l10n = context.l10n;
-    bool slotShown(int i) =>
-        destination == null ||
-        (i < labels.length && labels[i] == destination);
-    final showResiduals =
-        destination == null || destination == _kOtherPlaces;
-    return [
-      for (final (i, slot) in grouped.slots.indexed)
-        if (slotShown(i)) ...[
-          ..._bookingRowWidgets(slot, departureOnly: false),
-          if (i == grouped.slots.length - 1)
-            ..._bookingRowWidgets(slot, departureOnly: true),
-        ],
-      if (showResiduals) ...[
-        for (final todo in grouped.residual)
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 4),
-            child: BookingTodoCard(
-              todo: todo,
-              onBookedChanged: (v) => _setRowBooked(v, todo: todo),
-              onOpen: _openCallbackFor(todo),
-              openLabelOverride: _flightLegs.containsKey(todo.todoKey)
-                  ? l10n.tripFindFlights
-                  : null,
-              onEdit: todo.auto ? null : () => _editTodo(todo),
-              onDelete: todo.auto ? null : () => _deleteTodo(todo),
-            ),
-          ),
-        for (final a in grouped.residualStays)
-          _detailRowFor(stay: a, segment: null, detailOnly: true),
-        for (final s in grouped.residualSegments)
-          _detailRowFor(stay: null, segment: s, detailOnly: true),
-      ],
-    ];
-  }
-
-  /// Chips for the Bookings destination filter: the leg labels deduped in trip
-  /// order (a revisited city gets ONE chip covering both its runs — chips
-  /// select by label equality, and run-suffixed chips would leak the internal
-  /// `#2` key grammar), plus the canonical 'Other places' value when residual
-  /// bookings exist and no real 'Other places' leg already supplied it. Each
-  /// carries its own booked count from [bookingDestinationCounts].
-  ///
-  /// A stale selection (leg labels change when the itinerary is edited) is
-  /// clamped HERE rather than in either view body, because both scopes render
-  /// the strip and both need the same clamp before they filter their rows.
-  /// We're already in build, so this frame renders the clamped value (the
-  /// _focusedLegKey clamp idiom).
-  List<BookingDestination> _bookingsLensDestinations(
-    GroupedBookings grouped,
-    List<String> labels,
-  ) {
-    final l10n = context.l10n;
-    final hasResiduals = grouped.residual.isNotEmpty ||
-        grouped.residualStays.isNotEmpty ||
-        grouped.residualSegments.isNotEmpty;
-    final values = <String>[];
-    for (final l in labels) {
-      if (!values.contains(l)) values.add(l);
-    }
-    if (hasResiduals && !values.contains(_kOtherPlaces)) {
-      values.add(_kOtherPlaces);
-    }
-    if (_bookingsLensDestination != null &&
-        !values.contains(_bookingsLensDestination)) {
-      _bookingsLensDestination = null;
-    }
-    final counts =
-        bookingDestinationCounts(grouped, labels, otherKey: _kOtherPlaces);
-    return [
-      for (final v in values)
-        (
-          value: v,
-          label: v == _kOtherPlaces ? l10n.tripOtherBookings : v,
-          booked: counts[v]?.booked ?? 0,
-          total: counts[v]?.total ?? 0,
-        ),
-    ];
-  }
-
-  /// The Bookings view's one filter row — scope chip + destination strip —
-  /// rendered by BOTH scopes so toggling "Not booked yet" doesn't change the
-  /// shape of the chrome above the rows. The destination deliberately SURVIVES
-  /// that toggle: it is a different question ("where"), and clearing it would
-  /// silently widen what the traveler is looking at.
-  Widget _bookingsFilterBar(List<BookingDestination> destinations) =>
-      BookingFilterBar(
-        unbookedOnly: _itemFilter == 'unbooked',
-        onUnbookedOnlyChanged: (v) =>
-            setState(() => _itemFilter = v ? 'unbooked' : 'bookings'),
-        destinations: destinations,
-        selected: _bookingsLensDestination,
-        onSelected: (v) => setState(() => _bookingsLensDestination = v),
-      );
-
-  /// The 'unbooked' scope body: the same filter row as the all-bookings lens
-  /// above the flat left-to-book list. Never returns [] — the filter row is
-  /// the way out of an empty state, so it renders above both arms.
-  ///
-  /// The two empty arms say different true things. With no destination
-  /// chosen, an empty list means the TRIP is fully booked (the celebration).
-  /// With one chosen it means only that city is, so it gets its own line —
-  /// the trip-wide copy would be a claim this list cannot support.
-  List<Widget> _unbookedViewBody(
-    GroupedBookings grouped,
-    List<String> labels,
-  ) {
-    final l10n = context.l10n;
-    final destinations = _bookingsLensDestinations(grouped, labels);
-    final rows = _unbookedRows(grouped, labels,
-        destination: _bookingsLensDestination);
-    return [
-      _bookingsFilterBar(destinations),
-      if (rows.isEmpty && _bookingsLensDestination == null)
-        SizedBox(
-          height: 260,
-          child: EmptyState(
-            icon: Icons.celebration_outlined,
-            title: l10n.tripFilterAllBooked,
-            message: l10n.tripFilterAllBookedMessage,
-          ),
-        )
-      else if (rows.isEmpty)
-        SizedBox(
-          height: 120,
-          child: EmptyState(
-            icon: Icons.celebration_outlined,
-            title: l10n.tripBookingsAllBookedForDestination,
-            compact: true,
-          ),
-        )
-      else
-        ...rows,
-    ];
-  }
-
-  /// The 'bookings' lens body: the filter row above the flat all-bookings
-  /// list. Returns [] when the trip has no bookings at all so build can swap
-  /// in the lens empty state.
-  List<Widget> _bookingsLensBody(
-    GroupedBookings grouped,
-    List<String> labels,
-  ) {
-    final l10n = context.l10n;
-    final all = _allBookingRows(grouped, labels);
-    if (all.isEmpty) return const [];
-    // Builds the chips AND clamps a stale selection — so it runs before the
-    // rows below are filtered by that selection.
-    final destinations = _bookingsLensDestinations(grouped, labels);
-    final rows = _bookingsLensDestination == null
-        ? all
-        : _allBookingRows(grouped, labels,
-            destination: _bookingsLensDestination);
-    return [
-      _bookingsFilterBar(destinations),
-      if (rows.isEmpty)
-        SizedBox(
-          height: 120,
-          child: EmptyState(
-            icon: Icons.search_off,
-            title: l10n.tripBookingsLensNoneForDestination,
-            compact: true,
-          ),
-        )
-      else
-        ...rows,
-    ];
-  }
-
-  /// "+ Add booking" — the Bookings view's one add CTA (its "Add place").
-  /// A single menu fans out to the three record kinds so the itinerary tail
-  /// no longer needs a button trio. One MenuAnchor for both widths; only the
-  /// anchor swaps (labeled wide, icon-only narrow — same precedent and
-  /// reason as Add place below). Offline disables the anchor; each handler
-  /// re-guards via _guardOffline() for a menu left open across the flip.
-  Widget _addBookingMenu() {
-    final l10n = context.l10n;
-    return MenuAnchor(
-      menuChildren: [
-        MenuItemButton(
-          leadingIcon: const Icon(Icons.hotel_outlined, size: 18),
-          onPressed: _addStay,
-          child: Text(l10n.bookingsMenuStay),
-        ),
-        MenuItemButton(
-          leadingIcon: const Icon(Icons.route_outlined, size: 18),
-          onPressed: _addSegment,
-          child: Text(l10n.bookingsMenuTransport),
-        ),
-        MenuItemButton(
-          leadingIcon: const Icon(Icons.check_circle_outline, size: 18),
-          onPressed: _addBooking,
-          child: Text(l10n.bookingsMenuOther),
-        ),
-      ],
-      builder: (context, controller, _) {
-        void toggle() =>
-            controller.isOpen ? controller.close() : controller.open();
-        return _narrow
-            ? IconButton(
-                onPressed: _isOffline ? null : toggle,
-                tooltip: l10n.bookingsAddBooking,
-                visualDensity: VisualDensity.compact,
-                icon: const Icon(Icons.add, size: 20),
-              )
-            : TextButton.icon(
-                onPressed: _isOffline ? null : toggle,
-                style: TextButton.styleFrom(
-                  visualDensity: VisualDensity.compact,
-                ),
-                icon: const Icon(Icons.add, size: 18),
-                label: Text(l10n.bookingsAddBooking),
-              );
-      },
-    );
-  }
 
   /// One header view tab ("Itinerary" / "Bookings" + count pill). The
   /// selected tab doesn't take taps — re-tap is a no-op, which also keeps an
@@ -4182,1031 +2382,7 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     );
   }
 
-  /// Batches consecutive day-trip places (by town) under an indented
-  /// "Day trip · <town>" sub-header so nearby towns read as excursions from the
-  /// hub city rather than separate stops. Each contiguous batch (hub run or
-  /// day-trip batch) renders as its own [SliverReorderableList] so items can
-  /// be dragged inline within it; the within-city travel connector between
-  /// adjacent tiles is folded into the row below it and travels with it.
-  List<Widget> _dayTripSectionSlivers(
-      List<ItineraryItem> items, ThemeData theme) {
-    final slivers = <Widget>[];
-    var i = 0;
-    while (i < items.length) {
-      final dt = items[i].dayTripFrom?.trim();
-      if (dt != null && dt.isNotEmpty) {
-        final town = _cityOf(items[i]) ?? context.l10n.tripDayTripFallback;
-        slivers.add(_boxSliver([
-          _dayTripSubHeader(town, theme, _dayTripTravelLabel(items[i])),
-        ]));
-        final batch = <ItineraryItem>[];
-        while (i < items.length) {
-          final it = items[i];
-          final d = it.dayTripFrom?.trim();
-          if (d != null && d.isNotEmpty && _cityOf(it) == town) {
-            batch.add(it);
-            i++;
-          } else {
-            break;
-          }
-        }
-        slivers.add(_batchReorderableSliver(batch, 32, theme));
-      } else {
-        final batch = <ItineraryItem>[];
-        while (i < items.length &&
-            (items[i].dayTripFrom?.trim() ?? '').isEmpty) {
-          batch.add(items[i]);
-          i++;
-        }
-        slivers.add(_batchReorderableSliver(batch, 12, theme));
-      }
-    }
-    return slivers;
-  }
-
-  /// One contiguous batch of item tiles as an inline-draggable sliver. Drag is
-  /// confined to the batch, a subset of _sectionOf's boundary (items rendered
-  /// apart can't be dragged past each other — the menu's "Reorder section"
-  /// sheet still covers the full section).
-  Widget _batchReorderableSliver(
-      List<ItineraryItem> batch, double indent, ThemeData theme) {
-    // Narrow first: phones reorder via the kebab (Move up/down + Reorder
-    // section), so no handle is built at all — the same null-handle path
-    // offline/singleton batches already take.
-    final canDrag =
-        !_narrow && !_readOnly && !_isOffline && batch.length > 1;
-    return SliverReorderableList(
-      itemCount: batch.length,
-      // Unlike ReorderableListView, the sliver variant doesn't wrap the
-      // dragged proxy in a Material — without one the lifted ListTile throws.
-      proxyDecorator: (child, index, animation) => Material(
-        elevation: 4,
-        borderRadius: BorderRadius.circular(8),
-        child: child,
-      ),
-      onReorder: (oldIndex, newIndex) =>
-          _reorderBatchInline(batch, oldIndex, newIndex),
-      itemBuilder: (context, i) {
-        final item = batch[i];
-        final connector = i > 0
-            ? _travelConnector(batch[i - 1], item, indent, theme)
-            : null;
-        return Column(
-          key: ValueKey(item.id),
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            if (connector != null) connector,
-            _itemTile(
-              item,
-              indent,
-              theme,
-              dragHandle: canDrag
-                  ? ReorderableDragStartListener(
-                      index: i,
-                      child: Icon(Icons.drag_indicator,
-                          color: theme.colorScheme.onSurfaceVariant),
-                    )
-                  : null,
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  /// Persists an inline drag within one rendered batch. Optimistic: the new
-  /// order is spliced into the trip's item list in place so the drop doesn't
-  /// snap back while the request is in flight; the silent reload then
-  /// refreshes positions and travel connectors (or restores the server order
-  /// after a failure).
-  Future<void> _reorderBatchInline(
-      List<ItineraryItem> batch, int oldIndex, int newIndex) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final trip = _trip;
-    final items = trip?.items;
-    if (trip == null || items == null) return;
-    if (newIndex > oldIndex) newIndex--;
-    if (newIndex == oldIndex) return;
-    final newOrder = List.of(batch);
-    newOrder.insert(newIndex, newOrder.removeAt(oldIndex));
-    final batchIds = {for (final it in batch) it.id};
-    var next = 0;
-    final newItems = <ItineraryItem>[
-      for (final it in items)
-        if (batchIds.contains(it.id)) newOrder[next++] else it,
-    ];
-    setState(() {
-      items.setAll(0, newItems);
-      // NOTE: each item's `position` still holds its PRE-drag value until the
-      // silent reload returns — ItineraryItem is immutable by design. The city
-      // headers no longer care (TripDerivation indexes their date chips by leg,
-      // not by position), which is what kept a dragged frame from showing a
-      // neighbouring leg's dates.
-      // CONTRACT: any in-place mutation of a derivation input MUST bump the
-      // epoch. This setAll is the screen's ONE in-place write — the trip's
-      // item List keeps its identity, so [_derive]'s identity checks can't
-      // see the new order; the epoch bump is what invalidates the memo and
-      // lets this optimistic frame render the dragged order.
-      _itemOrderEpoch++;
-    });
-    try {
-      await ref
-          .read(tripsApiServiceProvider)
-          .reorderItineraryItems(trip.id, [for (final it in newItems) it.id]);
-      await _load(silent: true);
-    } catch (e) {
-      _showSnack(l10n.tripReorderFailed(friendlyError(l10n, e)));
-      await _load(silent: true);
-    }
-  }
-
-  /// The city header's name, with the repeat qualifier stacked beneath it when
-  /// another group renders the SAME label — a revisited city, or one a bad
-  /// section rewrite split in two. Without it the two headers read identically
-  /// and the duplicate looks like a rendering bug. [metaLine] stacks under
-  /// both on narrow (the date range — see [_cityHeaderMetaLine]).
-  ///
-  /// This is the ONE place that decides the stacked-label shape: everything
-  /// that drops below the name joins this Column, so the name's [Text] keeps
-  /// the header's outer Row as its nearest Row ancestor and the header tests
-  /// can still navigate that ancestry to pair a city with its dates.
-  ///
-  /// With neither a qualifier nor a meta line (every ordinary trip on a
-  /// desktop) this returns the bare [Text] the header has always used, so the
-  /// widget tree is unchanged.
-  Widget _cityHeaderLabel(
-      AppLocalizations l10n, CityGroup group, ThemeData theme,
-      {Widget? metaLine}) {
-    final label = Text(
-      groupLabelText(l10n, group.label),
-      maxLines: 1,
-      overflow: TextOverflow.ellipsis,
-      style: theme.textTheme.titleSmall?.copyWith(fontWeight: FontWeight.bold),
-    );
-    final qualifier = group.qualifier;
-    if (qualifier == null && metaLine == null) return label;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        label,
-        if (qualifier != null)
-          Text(
-            qualifier,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-            style: theme.textTheme.labelSmall
-                ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-          ),
-        if (metaLine != null) metaLine,
-      ],
-    );
-  }
-
-  /// The narrow city header's second line: "Aug 24 – Aug 27 · 3 nights".
-  ///
-  /// Carries no calendar icon — the pin above it already anchors the row, and
-  /// a second glyph on a phone is the noise this pass removes. The range and
-  /// the nights stay TWO Texts, as they have since the chip columns landed:
-  /// `tripLegNights` owns the leading "· ", so a zero-night leg drops the
-  /// middot structurally rather than by a conditional in the format string.
-  /// Both are Flexible so an extreme text scale ellipsizes inside the label
-  /// column instead of overflowing it.
-  Widget _cityHeaderMetaLine(CityGroup group, TextStyle? chipStyle) {
-    final nights = group.dateRange!.nights;
-    return MergeSemantics(
-      // One utterance, matching the reading order the chip had.
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Flexible(
-            child: Text(
-              group.dateRange!.range,
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-              style: chipStyle,
-            ),
-          ),
-          if (nights != null) ...[
-            const SizedBox(width: AppSpacing.xs),
-            Flexible(
-              child: Text(
-                nights,
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-                style: chipStyle,
-              ),
-            ),
-          ],
-        ],
-      ),
-    );
-  }
-
-  /// A small "↓ 12 min · 4.3 km" row shown between two consecutive itinerary
-  /// tiles, but only for within-city hops (same hub, truly adjacent in the
-  /// itinerary order). Returns null when it shouldn't render.
-  Widget? _travelConnector(ItineraryItem from, ItineraryItem to,
-      double indentLeft, ThemeData theme) {
-    if (to.position != from.position + 1) return null;
-    if (_hubOf(from) != _hubOf(to)) return null;
-    final timing = _travelByPos[from.position];
-    if (timing == null || timing.travelToNextMin <= 0) return null;
-
-    final km = timing.travelToNextKm;
-    final dist = km > 0 ? ' · ${km.toStringAsFixed(1)} km' : '';
-    final muted = theme.colorScheme.onSurfaceVariant;
-    final icon = km > 0 && km <= 1.2
-        ? Icons.directions_walk
-        : Icons.directions_car_outlined;
-    return Padding(
-      padding: EdgeInsets.only(left: indentLeft + 28, top: 2, bottom: 2),
-      child: Row(
-        children: [
-          Icon(icon, size: 14, color: muted),
-          const SizedBox(width: 6),
-          Text(
-            '${_fmtTravel(timing.travelToNextMin)}$dist',
-            style: theme.textTheme.bodySmall?.copyWith(color: muted),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// Total within-city travel time (minutes) across the consecutive legs of a
-  /// day's run.
-  int _runTravelMin(List<ItineraryItem> run) {
-    var total = 0;
-    for (var k = 0; k < run.length - 1; k++) {
-      final a = run[k];
-      final b = run[k + 1];
-      if (b.position == a.position + 1 && _hubOf(a) == _hubOf(b)) {
-        total += _travelByPos[a.position]?.travelToNextMin ?? 0;
-      }
-    }
-    return total;
-  }
-
   // The map's travel-time labels live in [TripDerivation.segmentLabels].
-
-  /// Travel time from the hub city to a day trip, e.g. "45 min from Paris",
-  /// taken from the already-computed leg into the day trip's first stop. Null
-  /// unless the preceding item is actually in the hub city (so a town-to-town
-  /// or cross-city leg is never mislabeled).
-  String? _dayTripTravelLabel(ItineraryItem first) {
-    final hub = first.dayTripFrom?.trim();
-    if (hub == null || hub.isEmpty) return null;
-    ItineraryItem? prev;
-    for (final it in _trip?.items ?? const <ItineraryItem>[]) {
-      if (it.position == first.position - 1) prev = it;
-    }
-    if (prev == null) return null;
-    final prevDayTrip = prev.dayTripFrom?.trim();
-    if (prevDayTrip != null && prevDayTrip.isNotEmpty) return null;
-    if (_hubOf(prev) != _hubOf(first)) return null;
-    final timing = _travelByPos[prev.position];
-    if (timing == null || timing.travelToNextMin <= 0) return null;
-    return context.l10n
-        .tripTravelFromHub(_fmtTravel(timing.travelToNextMin), hub);
-  }
-
-  /// Formats a travel duration: "45 min", "1h", or "1h 20m" (the shared
-  /// [fmtTravel], which the derivation's map labels also use).
-  String _fmtTravel(int min) => fmtTravel(context.l10n, min);
-
-  /// Day section header: shows the calendar date (day N -> startDate + (N-1))
-  /// when the trip start is known, otherwise falls back to "Day N". The opaque
-  /// Material keeps items from showing through while the header is pinned —
-  /// today's tint is alpha-blended onto the scaffold background (never a
-  /// translucent color) for the same reason. [headerKey] gives the Today
-  /// scroller a stable handle on the header's render box.
-  ///
-  /// [showMonth] is the caller's answer to "would dropping the month here lose
-  /// anything" — see the running-month rule in [_buildGroupItemSlivers]. It
-  /// only bites on narrow; a desktop row has the width to spell every date.
-  /// What a day row is called: the calendar date (day N = trip start + N-1)
-  /// when the trip has a start, else "Day N". ONE definition — the day header
-  /// and the empty-day placeholder below it must never disagree about which
-  /// date day N is, and they sit next to each other in the same list.
-  ///
-  /// The running-month rule lives HERE rather than at the header's call site so
-  /// the placeholder cannot drift from it: both rows ask the same function what
-  /// day N is called. Empty rows keep the default (month spelled out) because
-  /// they are emitted by flushEmptiesBefore, which does not carry the loop's
-  /// month state — a format difference on narrow, never a date difference.
-  String _dayHeaderLabel(int day, DateTime? tripStart,
-      {bool showMonth = true}) {
-    final date = tripStart?.add(Duration(days: day - 1));
-    if (date == null) return context.l10n.tripDayN(day);
-    return _narrow && !showMonth ? weekdayDay(date) : _fmtDayHeader(date);
-  }
-  Widget _daySubHeader(
-      int day,
-      DateTime? tripStart,
-      ThemeData theme,
-      bool collapsed,
-      int travelMin,
-      VoidCallback onTap,
-      VoidCallback? onRefine,
-      {Key? headerKey,
-      bool isToday = false,
-      bool showMonth = true}) {
-    final l10n = context.l10n;
-    final label = _dayHeaderLabel(day, tripStart, showMonth: showMonth);
-    final muted = theme.colorScheme.onSurfaceVariant;
-    final header = HoverReveal(
-      builder: (context, revealed) => Material(
-        key: headerKey,
-        color: isToday
-            ? Color.alphaBlend(
-                theme.colorScheme.primary.withValues(alpha: 0.06),
-                theme.scaffoldBackgroundColor)
-            : theme.scaffoldBackgroundColor,
-        child: InkWell(
-          onTap: onTap,
-          child: Padding(
-            // Narrow drops the calendar glyph and indents to where the city
-            // name (and its dates) start instead: two calendar icons stacked
-            // two rows apart said less than one clean indent column does, and
-            // the day is subordinate to the city, so it should read that way.
-            padding: _narrow
-                ? const EdgeInsets.fromLTRB(24, 8, 16, 2)
-                : const EdgeInsets.fromLTRB(16, 12, 16, 2),
-            child: Row(
-              children: [
-                if (!_narrow) ...[
-                  Icon(Icons.today, size: 16, color: theme.colorScheme.primary),
-                  const SizedBox(width: 6),
-                ],
-                Expanded(
-                  child: Row(
-                    children: [
-                      Flexible(
-                        child: Text(
-                          label,
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                          style: theme.textTheme.labelLarge?.copyWith(
-                            color: theme.colorScheme.primary,
-                            fontWeight: FontWeight.w700,
-                          ),
-                        ),
-                      ),
-                      if (isToday) ...[
-                        const SizedBox(width: 8),
-                        StatusPill.custom(
-                          label: l10n.tripToday,
-                          background: theme.colorScheme.primary
-                              .withValues(alpha: 0.15),
-                          foreground: theme.colorScheme.primary,
-                        ),
-                      ],
-                    ],
-                  ),
-                ),
-                // Narrow drops the day travel-total: it is fixed-width (no
-                // ellipsis) and fights the date label, and the per-leg
-                // connectors below carry the same information.
-                if (travelMin > 0 && !_narrow) ...[
-                  Icon(Icons.directions_car_outlined, size: 14, color: muted),
-                  const SizedBox(width: 4),
-                  Text(
-                    l10n.tripTravelTotal(_fmtTravel(travelMin)),
-                    style: theme.textTheme.bodySmall?.copyWith(color: muted),
-                  ),
-                  const SizedBox(width: 8),
-                ],
-                // Hover-revealed on pointer devices (always visible on
-                // touch) to keep the resting row quiet.
-                if (onRefine != null)
-                  HoverRevealed(
-                    revealed: revealed,
-                    child: IconButton(
-                      icon: const Icon(Icons.auto_awesome, size: 16),
-                      tooltip: l10n.tripRefineThisDay,
-                      visualDensity: VisualDensity.compact,
-                      color: theme.colorScheme.primary,
-                      onPressed: onRefine,
-                    ),
-                  ),
-                Icon(
-                  collapsed ? Icons.chevron_right : Icons.expand_more,
-                  size: 18,
-                  color: theme.colorScheme.primary,
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-    // Repaint-isolate like _cityHeader: open days pin their header over the
-    // scrolling list; hover/InkWell repaints must stay in this layer.
-    return RepaintBoundary(child: header);
-  }
-
-  /// A day inside a city leg that carries nothing — the state the two-pass
-  /// planner makes normal (specs/shape-before-schedule). Before this, such a
-  /// day rendered NOTHING at all: day sub-headers are derived from the items
-  /// grouped under a leg, so a Rome leg spanning Sep 1-5 with places on days 1
-  /// and 4 gave no sign that 2 and 3 existed.
-  ///
-  /// Deliberately NOT a pinned header and NOT collapsible: there is no body to
-  /// pin over, and a zero-body pinned sliver poisons the surrounding
-  /// MultiSliver's paint origin. Quieter than a planned day — muted glyph and
-  /// label — because an open day is an invitation, not a plan.
-  ///
-  /// [onPlan] is null for viewers and offline, matching the day header's own
-  /// refine gate; the row then renders as a plain statement of the gap.
-  Widget _emptyDayRow(int day, String dayKey, DateTime? tripStart,
-      ThemeData theme,
-      {bool isToday = false, VoidCallback? onPlan}) {
-    final l10n = context.l10n;
-    final muted = theme.colorScheme.onSurfaceVariant;
-    return RepaintBoundary(
-      key: ValueKey('unplanned-day:$dayKey'),
-      child: Material(
-        // The same GlobalKey registry the day headers use, so day-jump can
-        // scroll to an open day exactly as it scrolls to a planned one.
-        key: _dayHeaderKeys.putIfAbsent(dayKey, GlobalKey.new),
-        color: isToday
-            ? Color.alphaBlend(
-                theme.colorScheme.primary.withValues(alpha: 0.06),
-                theme.scaffoldBackgroundColor)
-            : theme.scaffoldBackgroundColor,
-        child: InkWell(
-          onTap: onPlan,
-          child: Padding(
-            padding: const EdgeInsets.fromLTRB(16, 12, 16, 2),
-            child: Row(
-              children: [
-                Icon(Icons.today, size: 16, color: muted),
-                const SizedBox(width: 8),
-                Flexible(
-                  child: Text(
-                    _dayHeaderLabel(day, tripStart),
-                    style: theme.textTheme.labelLarge?.copyWith(color: muted),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Flexible(
-                  child: Text(
-                    l10n.tripDayNothingPlanned,
-                    style: theme.textTheme.bodySmall?.copyWith(color: muted),
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-                const Spacer(),
-                // A label, not a nested button: the row's own InkWell is the
-                // tap target, so there is one semantics node here rather than
-                // two offering the same action.
-                if (onPlan != null) ...[
-                  Icon(Icons.auto_awesome,
-                      size: 16, color: theme.colorScheme.primary),
-                  const SizedBox(width: 4),
-                  // Ellipsizing like the two labels beside it: Spanish at 1.3x
-                  // on a 360px viewport overflowed this row by 86px while this
-                  // was the one unflexible child (specs/booking-shortlist's
-                  // narrow-Spanish test is what caught it).
-                  Flexible(
-                    child: Text(
-                      l10n.tripPlanThisDay,
-                      style: theme.textTheme.labelMedium
-                          ?.copyWith(color: theme.colorScheme.primary),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                ],
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  /// Per-day weather chip (specs/weather-in-itinerary): hi/lo temp + a
-  /// condition glyph derived from rain, rendered as a non-pinned content row
-  /// under the day header (same indent/typography as [_tonightCaption]).
-  /// Forecast days append the rain chance; historical reports (trip beyond the
-  /// 16-day horizon) are labeled "typical", never "forecast".
-  Widget _weatherChip(ThemeData theme, WeatherDay day, bool historical) {
-    final muted = theme.colorScheme.onSurfaceVariant;
-    final (icon, iconColor) = _weatherGlyph(theme, day);
-    final hi = day.tempMaxC.round();
-    final lo = day.tempMinC.round();
-    final parts = <String>['$hi° / $lo°'];
-    if (!historical &&
-        day.precipProbability != null &&
-        day.precipProbability! >= rainChanceCaptionPct) {
-      parts.add(context.l10n.tripRainChance(day.precipProbability!));
-    }
-    return Padding(
-      // Matches the Tonight caption indent (20 left, 6 top) so the chip reads
-      // as a sub-row of the day header.
-      padding: const EdgeInsets.fromLTRB(20, 6, 16, 0),
-      child: Row(
-        children: [
-          Icon(icon, size: 15, color: iconColor),
-          const SizedBox(width: 6),
-          Flexible(
-            child: Text(
-              parts.join('  ·  '),
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-              style: theme.textTheme.bodySmall?.copyWith(color: muted),
-            ),
-          ),
-          if (historical) ...[
-            const SizedBox(width: 8),
-            Flexible(
-              child: Text(
-                context.l10n.tripTypicalForDates,
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: muted,
-                  fontStyle: FontStyle.italic,
-                ),
-              ),
-            ),
-          ],
-        ],
-      ),
-    );
-  }
-
-  /// Condition glyph + color from precipitation: sunny / cloud / umbrella by
-  /// [rainLevel] — the client's one rain-threshold definition, shared with
-  /// the what-to-wear recommendations (utils/clothing_recs.dart).
-  (IconData, Color) _weatherGlyph(ThemeData theme, WeatherDay day) {
-    final scheme = theme.colorScheme;
-    return switch (rainLevel(day)) {
-      RainLevel.likely => (Icons.umbrella, scheme.primary),
-      RainLevel.some => (Icons.cloud, scheme.onSurfaceVariant),
-      RainLevel.none => (Icons.wb_sunny, Colors.amber.shade700),
-    };
-  }
-
-  /// "Tonight: <stay>" caption for today's day section (specs/happening-now
-  /// PR 2): where the traveler sleeps tonight, checkout-exclusively. Returns
-  /// null when no covering stay has a non-empty name — no filler row.
-  Widget? _tonightCaption(ThemeData theme, List<Accommodation> stays) {
-    final names = stays
-        .map((a) => a.name.trim())
-        .where((n) => n.isNotEmpty)
-        .toList();
-    if (names.isEmpty) return null;
-    return Padding(
-      // Matches the _dayTripSubHeader indent (20 left) so the caption reads
-      // as a sub-row of the day, tucked tight under the header (6 top).
-      padding: const EdgeInsets.fromLTRB(20, 6, 16, 0),
-      child: Row(
-        children: [
-          Icon(Icons.hotel, size: 14, color: theme.colorScheme.primary),
-          const SizedBox(width: 6),
-          Expanded(
-            child: Text(
-              context.l10n.tripTonight(names.join(', ')),
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-              style: theme.textTheme.bodySmall
-                  ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _dayTripSubHeader(String town, ThemeData theme, String? travelLabel) =>
-      Padding(
-        padding: const EdgeInsets.fromLTRB(20, 8, 16, 0),
-        child: Row(
-          children: [
-            Icon(Icons.directions_bus,
-                size: 16, color: theme.colorScheme.secondary),
-            const SizedBox(width: 6),
-            Expanded(
-              child: Text(
-                context.l10n.tripDayTripTo(town),
-                style: theme.textTheme.labelLarge?.copyWith(
-                  color: theme.colorScheme.secondary,
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-            ),
-            if (travelLabel != null) ...[
-              Icon(Icons.directions_car_outlined,
-                  size: 14, color: theme.colorScheme.onSurfaceVariant),
-              const SizedBox(width: 4),
-              Text(
-                travelLabel,
-                style: theme.textTheme.bodySmall
-                    ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-              ),
-            ],
-          ],
-        ),
-      );
-
-  /// Opens the add-to-trip picker for a browsed place (local rec / event) with
-  /// this trip preselected, then refreshes in place when the add landed here.
-  Future<void> _addToTrip(AddToTripPayload payload) async {
-    final added =
-        await showAddToTripSheet(context, payload, currentTripId: widget.tripId);
-    // _refresh(), not a bare silent _load(): it serializes with any reload the
-    // refine panel has in flight, so a pre-add snapshot can't land after us
-    // and momentarily erase the just-added item. Armed like the chat path —
-    // an add the traveler just made must be visible, folded city or not.
-    if (added != null && added.id == widget.tripId) {
-      _armRevealOfNewItems();
-      _refresh();
-    }
-  }
-
-  Widget _itemTile(ItineraryItem item, double indentLeft, ThemeData theme,
-          {Widget? dragHandle}) =>
-      // Selection is notifier-driven (see _selectedPosition): each tile
-      // listens itself, so selecting a place rebuilds only the visible
-      // tiles rather than the whole screen.
-      ValueListenableBuilder<int?>(
-        valueListenable: _selectedPosition,
-        builder: (context, selectedPos, _) =>
-            // Secondary controls (maps link, kebab, drag handle) reveal on
-            // hover: opacity-only, so layout never shifts, drag geometry
-            // stays stable for the reorderable list, and touch devices (no
-            // mouse) keep them permanently visible. The time-of-day chip is
-            // content, not a control — it stays.
-            HoverReveal(
-        builder: (context, revealed) => Padding(
-          padding: EdgeInsets.only(left: indentLeft),
-          child: ListTile(
-          leading: _itemLeading(item.category, item.position),
-          title: Text(item.name,
-              maxLines: 2, overflow: TextOverflow.ellipsis),
-          subtitle: (item.address != null || item.localSourceName != null)
-              ? Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    if (item.address != null)
-                      Text(item.address!,
-                          maxLines: 1, overflow: TextOverflow.ellipsis),
-                    // The local-source credit line: who vouched for this place
-                    // (snapshot; shown for agent- and browse-added items alike).
-                    if (item.localSourceName != null)
-                      Row(
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          Icon(Icons.verified,
-                              size: 13, color: AppColors.toolLocal),
-                          const SizedBox(width: 4),
-                          Flexible(
-                            child: Text(
-                              context.l10n
-                                  .tripRecommendedBy(item.localSourceName!),
-                              style: theme.textTheme.labelSmall?.copyWith(
-                                color: AppColors.toolLocal,
-                                fontWeight: FontWeight.w600,
-                              ),
-                              maxLines: 1,
-                              overflow: TextOverflow.ellipsis,
-                            ),
-                          ),
-                        ],
-                      ),
-                  ],
-                )
-              : null,
-          trailing: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              if (item.timeOfDay != null)
-                _TimeOfDayChip(
-                    timeOfDay: item.timeOfDay!, compact: _narrow),
-              HoverRevealed(
-                revealed: revealed,
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    // On narrow, Maps lives in the kebab and this button
-                    // goes — EXCEPT for viewers, who have no kebab (gate
-                    // below) and would otherwise lose their only maps
-                    // access. Role-only on purpose: offline editors keep
-                    // the kebab, so they keep Maps through it.
-                    if (!_narrow || _readOnly)
-                      IconButton(
-                        icon: const Icon(Icons.map_outlined),
-                        tooltip: context.l10n.tripOpenInGoogleMaps,
-                        onPressed: () => _launch(_mapsUrl(item)),
-                      ),
-                    if (!_readOnly) _itemMenu(item),
-                    if (dragHandle != null) dragHandle,
-                  ],
-                ),
-              ),
-            ],
-          ),
-          selected: selectedPos == item.position,
-          selectedTileColor:
-              theme.colorScheme.primary.withValues(alpha: 0.08),
-          // The map is pinned and always on screen, so tapping an item only
-          // needs to update the selection; the notifier rebuilds the map
-          // card (TripMap recenters) and the visible tiles — no setState.
-          onTap: () => _selectedPosition.value = item.position,
-        ),
-        ),
-        ),
-      );
-
-  /// Per-item actions: edit, move within its section, delete (with undo).
-  /// Move targets the neighbor in itinerary order but only within the same
-  /// day + hub + day-trip batch, so an item can never silently jump across a
-  /// section boundary — cross-day moves go through the edit sheet instead.
-  Widget _itemMenu(ItineraryItem item) {
-    final l10n = context.l10n;
-    return PopupMenuButton<String>(
-      icon: const Icon(Icons.more_vert),
-      tooltip: l10n.tripPlaceActions,
-      onSelected: (action) {
-        switch (action) {
-          case 'edit':
-            _editItem(item);
-          case 'maps':
-            _launch(_mapsUrl(item));
-          case 'up':
-            _moveItem(item, -1);
-          case 'down':
-            _moveItem(item, 1);
-          case 'reorder':
-            _reorderSection(item);
-          case 'gcal':
-            _addItemToGoogleCalendar(item);
-          case 'ics':
-            _addItemToAppleCalendar(item);
-          case 'delete':
-            _deleteItem(item);
-        }
-      },
-      itemBuilder: (context) {
-        // Computed lazily — only when the menu actually opens. At the old
-        // call-time spot these three ran O(items) work per rendered tile per
-        // build (specs/perf-program, Wave 4 PR1).
-        final canUp = _moveNeighbor(item, -1) != null;
-        final canDown = _moveNeighbor(item, 1) != null;
-        final calendarRange = itemCalendarRange(_trip?.startDate, item.day);
-        return [
-        PopupMenuItem(
-          value: 'edit',
-          child: ListTile(
-            leading: const Icon(Icons.edit_outlined),
-            title: Text(l10n.tripEdit),
-            contentPadding: EdgeInsets.zero,
-          ),
-        ),
-        // All widths, not just narrow (where the standalone button is gone):
-        // one menu shape everywhere, and the desktop hover icon is invisible
-        // at rest, so a labeled entry helps discovery.
-        PopupMenuItem(
-          value: 'maps',
-          child: ListTile(
-            leading: const Icon(Icons.map_outlined),
-            title: Text(l10n.tripOpenInGoogleMaps),
-            contentPadding: EdgeInsets.zero,
-          ),
-        ),
-        if (canUp)
-          PopupMenuItem(
-            value: 'up',
-            child: ListTile(
-              leading: const Icon(Icons.arrow_upward),
-              title: Text(l10n.tripMoveUp),
-              contentPadding: EdgeInsets.zero,
-            ),
-          ),
-        if (canDown)
-          PopupMenuItem(
-            value: 'down',
-            child: ListTile(
-              leading: const Icon(Icons.arrow_downward),
-              title: Text(l10n.tripMoveDown),
-              contentPadding: EdgeInsets.zero,
-            ),
-          ),
-        if (_sectionOf(item).length > 2)
-          PopupMenuItem(
-            value: 'reorder',
-            child: ListTile(
-              leading: const Icon(Icons.drag_indicator),
-              title: Text(l10n.tripReorderSection),
-              contentPadding: EdgeInsets.zero,
-            ),
-          ),
-        if (calendarRange != null) ...[
-          PopupMenuItem(
-            value: 'gcal',
-            child: ListTile(
-              leading: const Icon(Icons.event_outlined),
-              title: Text(l10n.tripAddToGoogleCalendar),
-              contentPadding: EdgeInsets.zero,
-            ),
-          ),
-          PopupMenuItem(
-            value: 'ics',
-            enabled: !_isOffline,
-            child: ListTile(
-              leading: const Icon(Icons.event_available_outlined),
-              title: Text(l10n.tripAddToAppleCalendar),
-              contentPadding: EdgeInsets.zero,
-            ),
-          ),
-        ],
-        PopupMenuItem(
-          value: 'delete',
-          child: ListTile(
-            leading: const Icon(Icons.delete_outline),
-            title: Text(l10n.tripRemove),
-            contentPadding: EdgeInsets.zero,
-          ),
-        ),
-        ];
-      },
-    );
-  }
-
-  /// The item this one would swap with when moved by [delta] (-1 up, +1 down),
-  /// or null when the move would cross a day/hub/day-trip boundary.
-  ItineraryItem? _moveNeighbor(ItineraryItem item, int delta) {
-    final items = _trip?.items;
-    if (items == null) return null;
-    final idx = items.indexWhere((i) => i.id == item.id);
-    if (idx < 0) return null;
-    final ni = idx + delta;
-    if (ni < 0 || ni >= items.length) return null;
-    final other = items[ni];
-    if (other.day != item.day) return null;
-    if (_hubOf(other) != _hubOf(item)) return null;
-    if ((other.dayTripFrom ?? '').trim() != (item.dayTripFrom ?? '').trim()) {
-      return null;
-    }
-    return other;
-  }
-
-  /// All items sharing [item]'s day + hub + day-trip batch, in itinerary
-  /// order — the same boundary _moveNeighbor enforces, so drag reordering
-  /// can never move an item across a section either.
-  List<ItineraryItem> _sectionOf(ItineraryItem item) {
-    final items = _trip?.items ?? const <ItineraryItem>[];
-    return [
-      for (final i in items)
-        if (i.day == item.day &&
-            _hubOf(i) == _hubOf(item) &&
-            (i.dayTripFrom ?? '').trim() == (item.dayTripFrom ?? '').trim())
-          i,
-    ];
-  }
-
-  /// Drag-and-drop reorder for one section (specs/itinerary-item-editing
-  /// follow-up). The sheet reorders locally; Save maps the section's new
-  /// order back onto the full item-id permutation and submits it through the
-  /// same PUT /items/order path as Move up/down.
-  Future<void> _reorderSection(ItineraryItem item) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final trip = _trip;
-    if (trip == null) return;
-    final section = _sectionOf(item);
-    if (section.length < 2) return;
-
-    final newOrder = await showModalBottomSheet<List<ItineraryItem>>(
-      context: context,
-      isScrollControlled: true,
-      useSafeArea: true,
-      builder: (context) => _ReorderSectionSheet(items: section),
-    );
-    if (newOrder == null || !mounted) return;
-
-    // Splice the section's new order into the full ordering: walk the trip's
-    // items and replace each section member slot with the next reordered id.
-    final sectionIds = section.map((i) => i.id).toSet();
-    var next = 0;
-    final ids = <String>[
-      for (final i in trip.items ?? const <ItineraryItem>[])
-        if (sectionIds.contains(i.id)) newOrder[next++].id else i.id,
-    ];
-    try {
-      await ref
-          .read(tripsApiServiceProvider)
-          .reorderItineraryItems(trip.id, ids);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripReorderFailed(friendlyError(l10n, e)));
-      await _load();
-    }
-  }
-
-  Future<void> _moveItem(ItineraryItem item, int delta) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final trip = _trip;
-    final other = _moveNeighbor(item, delta);
-    if (trip == null || other == null) return;
-    final ids = (trip.items ?? const <ItineraryItem>[])
-        .map((i) => i.id)
-        .toList();
-    final a = ids.indexOf(item.id);
-    final b = ids.indexOf(other.id);
-    ids[a] = other.id;
-    ids[b] = item.id;
-    try {
-      await ref
-          .read(tripsApiServiceProvider)
-          .reorderItineraryItems(trip.id, ids);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripReorderFailed(friendlyError(l10n, e)));
-      await _load();
-    }
-  }
-
-  Future<void> _deleteItem(ItineraryItem item) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final trip = _trip;
-    if (trip == null) return;
-    try {
-      await ref
-          .read(tripsApiServiceProvider)
-          .deleteItineraryItem(trip.id, item.id);
-    } catch (e) {
-      _showSnack(l10n.tripRemoveItemFailed(item.name, friendlyError(l10n, e)));
-      return;
-    }
-    await _load();
-    if (!mounted) return;
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(l10n.tripRemovedItem(item.name)),
-        action: SnackBarAction(
-          label: l10n.tripUndo,
-          onPressed: () => _undoDelete(trip.id, item),
-        ),
-      ),
-    );
-  }
-
-  /// Undo = re-add through the normal add endpoint; the server slots the item
-  /// back at the end of its day, which is close enough to where it was.
-  Future<void> _undoDelete(String tripId, ItineraryItem item) async {
-    final l10n = context.l10n;
-    final body = <String, dynamic>{
-      'name': item.name,
-      if (item.placeId != null) 'place_id': item.placeId,
-      if (item.address != null) 'address': item.address,
-      if (item.latitude != 0 || item.longitude != 0) ...{
-        'latitude': item.latitude,
-        'longitude': item.longitude,
-      },
-      if (item.category != null) 'category': item.category,
-      if (item.timeOfDay != null) 'time_of_day': item.timeOfDay,
-      if (item.city != null) 'city': item.city,
-      if (item.dayTripFrom != null) 'day_trip_from': item.dayTripFrom,
-      if (item.day != null) 'day': item.day,
-    };
-    try {
-      await ref.read(tripsApiServiceProvider).addItineraryItem(tripId, body);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripRestoreItemFailed(item.name, friendlyError(l10n, e)));
-    }
-  }
-
-  Future<void> _editItem(ItineraryItem item) async {
-    if (_guardOffline()) return;
-    final l10n = context.l10n;
-    final changes = await showModalBottomSheet<Map<String, dynamic>>(
-      context: context,
-      isScrollControlled: true,
-      builder: (_) => _EditItineraryItemSheet(item: item),
-    );
-    if (changes == null || changes.isEmpty) return;
-    final trip = _trip;
-    if (trip == null) return;
-    try {
-      await ref
-          .read(tripsApiServiceProvider)
-          .updateItineraryItem(trip.id, item.id, changes);
-      await _load();
-    } catch (e) {
-      _showSnack(l10n.tripUpdateItemFailed(item.name, friendlyError(l10n, e)));
-    }
-  }
 
   // ---- Trip Health one-tap fixes ------------------------------------------
 
@@ -5411,55 +2587,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     );
   }
 
-  /// The Next Step card area (specs/next-step-cta), its own Consumer so review
-  /// refetches repaint the card alone — the `_healthAppBarAction` isolation
-  /// pattern. Hidden for viewers, while the review has no value (first load /
-  /// error / viewer 404), and for past trips (the server omits the step). The
-  /// all_set celebration is session-scoped via [_hadNextStep] /
-  /// [_allSetDismissed].
-  Widget _nextStepArea(Trip trip) {
-    if (_readOnly) return const SizedBox.shrink();
-    return Consumer(
-      builder: (context, ref, _) {
-        final review =
-            ref.watch(tripReviewProvider(TripReviewKey(trip.id))).valueOrNull;
-        final step = review?.nextStep;
-        if (step == null) return const SizedBox.shrink();
-        final allSet = step.kind == 'all_set';
-        if (!allSet && !_hadNextStep) {
-          // Record "this session saw a real step" post-frame (no setState in
-          // build); the guard makes it one-shot.
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            if (mounted && !_hadNextStep) setState(() => _hadNextStep = true);
-          });
-        }
-        if (allSet && (!_hadNextStep || _allSetDismissed)) {
-          return const SizedBox.shrink();
-        }
-        final progress = review?.planProgress;
-        return NextStepCard(
-          step: step,
-          progress: progress,
-          compact: _narrow,
-          enabled: !_isOffline,
-          // Same lookup the tap performs, so the label can never promise a
-          // handoff the action won't make (specs/next-step-cta).
-          transportHandsOff: _transportHandsOff(step),
-          onPrimary: allSet ? null : () => _onNextStepAction(trip, step),
-          onViewAll: () => _openHealthSheet(trip),
-          // No ladder on the wire (older server, cached response) => no
-          // affordance, rather than an entry point onto an empty sheet.
-          onViewProgress: progress != null && progress.phases.isNotEmpty
-              ? () => showPlanProgressSheet(context,
-                  progress: progress, currentStep: step)
-              : null,
-          onDismiss:
-              allSet ? () => setState(() => _allSetDismissed = true) : null,
-        );
-      },
-    );
-  }
-
   /// Routes the Next Step card's primary action: planning steps seed the trip
   /// chat with the server-built prompt; mechanical steps jump to the matching
   /// control directly (mixed-actions decision, specs/next-step-cta).
@@ -5590,19 +2717,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     if (saved) await _afterFix();
   }
 
-  /// Google Maps deep link for a place: prefer place_id, then coordinates, then a
-  /// name/address text search.
-  String _mapsUrl(ItineraryItem it) {
-    const base = 'https://www.google.com/maps/search/?api=1';
-    if (it.placeId != null && it.placeId!.isNotEmpty) {
-      return '$base&query=${Uri.encodeComponent(it.name)}&query_place_id=${it.placeId}';
-    }
-    if (it.latitude != 0 || it.longitude != 0) {
-      return '$base&query=${it.latitude},${it.longitude}';
-    }
-    return '$base&query=${Uri.encodeComponent('${it.name} ${it.address ?? ''}'.trim())}';
-  }
-
   // The per-position date chips live in [TripDerivation.locationDates]; the
   // leg date-range derivation itself (raw + visible) stays in
   // utils/leg_ranges.dart (specs/trip-dates-truth stage 0a) — one testable
@@ -5610,59 +2724,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   // behind the server legs payload.
 
   /// Coarse relative timestamp for the "Updated by X" line.
-  String _relativeTime(String iso) {
-    final l10n = context.l10n;
-    final t = DateTime.tryParse(iso);
-    if (t == null) return l10n.tripTimeRecently;
-    final d = DateTime.now().difference(t.toLocal());
-    if (d.inMinutes < 1) return l10n.tripTimeJustNow;
-    if (d.inMinutes < 60) return l10n.tripTimeMinutesAgo(d.inMinutes);
-    if (d.inHours < 24) return l10n.tripTimeHoursAgo(d.inHours);
-    return l10n.tripTimeDaysAgo(d.inDays);
-  }
-
-  /// Day-header date, e.g. "Wed, Jul 15" (weekday + month + day); Spanish
-  /// reorders to "mié, 15 jul" on its own.
-  String _fmtDayHeader(DateTime d) => mmmed().format(d);
-
-  /// The itinerary's trailing "Other bookings" area — the home for
-  /// everything the retired Bookings section held that has no city slot:
-  /// residual todos (custom bookings, stale autos) and confirmed records
-  /// that matched no city. Content-only — the add actions live in the
-  /// Bookings view's header button (_addBookingMenu).
-  Widget? _otherBookingsArea(
-      ThemeData theme,
-      ({
-        List<BookingTodo> residual,
-        List<Accommodation> residualStays,
-        List<TripSegment> residualSegments,
-      }) other) {
-    final l10n = context.l10n;
-    final hasContent = other.residual.isNotEmpty ||
-        other.residualStays.isNotEmpty ||
-        other.residualSegments.isNotEmpty;
-    if (!hasContent) return null;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Padding(
-          padding:
-              const EdgeInsets.only(top: AppSpacing.sm, bottom: AppSpacing.xs),
-          child: Text(
-            l10n.tripOtherBookings,
-            style: theme.textTheme.titleSmall
-                ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-          ),
-        ),
-        if (other.residual.isNotEmpty)
-          _residualBookingsList(other.residual, theme),
-        for (final a in other.residualStays)
-          _detailRowFor(stay: a, segment: null, detailOnly: true),
-        for (final s in other.residualSegments)
-          _detailRowFor(stay: null, segment: s, detailOnly: true),
-      ],
-    );
-  }
 
   /// Per-leg what-to-wear derivations for the wear/pack sheet
   /// (specs/what-to-wear). Queried AND displayed on visibleLegRanges — the
@@ -6033,358 +3094,10 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
     );
   }
 
-  /// The trip's hero header: title (+ rename), date/status chips, a Refine
-  /// button, and a collapsible overview.
-  Widget _buildHeaderCard(Trip trip, ThemeData theme) {
-    final l10n = context.l10n;
-    final overview = _overviewText(trip);
-    final hasDates = trip.startDate != null && trip.endDate != null;
-    // Deliberately card-less: the app bar already carries the title, so this
-    // block is a compact anchor (rename affordance + meta chips + context
-    // line + clamped overview), not a hero panel.
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Expanded(
-              child: Text(
-                _displayTitle(trip),
-                // Same string as the app bar directly above, so same face —
-                // but keeping titleLarge's size, because this block is a
-                // compact anchor and headlineSmall would inflate it into the
-                // hero panel the comment above rules out.
-                style: theme.textTheme.titleLarge?.copyWith(
-                  fontFamily: AppFonts.display,
-                  fontWeight: FontWeight.w400,
-                ),
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
-            if (trip.canEdit)
-              IconButton(
-                icon: const Icon(Icons.edit, size: 20),
-                tooltip: l10n.tripEditDetails,
-                onPressed: _isOffline ? null : _editTripDetails,
-              ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Wrap(
-          spacing: 8,
-          runSpacing: 8,
-          crossAxisAlignment: WrapCrossAlignment.center,
-          children: [
-            ActionChip(
-              avatar: const Icon(Icons.event, size: 16),
-              // Humanized short range ("Jul 20 – Jul 27") at every width, and
-              // localized with it (tripDateRange goes through the app locale,
-              // so Spanish reads "20 jul – 27 jul"). Narrow got this first,
-              // because the raw ISO pair is ~200px and forced the meta row to
-              // wrap; wide simply kept the machine form it started with, which
-              // left the SAME trip reading "2026-07-21 → 2026-07-22" on desktop
-              // and "Jul 21 – Jul 22" on a phone. Width is not a reason to
-              // show a different date format — and the layout that has more
-              // room was the one showing the denser string.
-              //
-              // The ISO pair survives only as the unparseable-date fallback:
-              // tripDateRange returns null there, and echoing back whatever is
-              // stored beats an empty chip on a trip whose dates are the thing
-              // you came to fix.
-              label: Text(hasDates
-                  ? (tripDateRange(trip.startDate, trip.endDate) ??
-                      '${trip.startDate} → ${trip.endDate}')
-                  : l10n.tripAddDates),
-              onPressed: (_isOffline || !trip.canEdit) ? null : _editDates,
-            ),
-            // The draft/planned status pill is gone with the status concept
-            // itself (specs/retire-trip-status) — dates carry the state.
-            // The trip-wide travel-mode pill is gone: transport mode is
-            // per-leg now, picked directly on each transport row (the
-            // _ModeMenu in BookingTodoRow). trips.travel_mode remains the
-            // AI-facing trip default behind _groundModeOf.
-            // Refine entry, demoted from a full-width banner to a peer of the
-            // meta chips. Same canEdit gate as before, so editor
-            // collaborators keep their spec-mandated entry point
-            // (specs/collaborator-refine); the per-city/day sparkles and the
-            // chat FAB are unchanged. On narrow this moves to the app bar.
-            if (trip.canEdit && !_narrow)
-              FilledButton.tonalIcon(
-                // Chat/refine needs the network — disabled while offline.
-                onPressed: _isOffline
-                    ? null
-                    : () => _openRefine(trip, const RefineTarget.trip()),
-                style: FilledButton.styleFrom(
-                    visualDensity: VisualDensity.compact),
-                icon: const Icon(Icons.auto_awesome, size: 16),
-                label: Text(l10n.tripRefineWithAI),
-              ),
-          ],
-        ),
-        // Muted context line: collaborator standing and/or "Updated by
-        // Maria · 2m ago" (the server omits self-attribution). Either part
-        // can stand alone.
-        if (!trip.isOwner || trip.updatedByName != null) ...[
-          const SizedBox(height: 8),
-          Wrap(
-            spacing: 12,
-            runSpacing: 4,
-            crossAxisAlignment: WrapCrossAlignment.center,
-            children: [
-              if (!trip.isOwner)
-                Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Icon(
-                        trip.canEdit
-                            ? Icons.group_outlined
-                            : Icons.visibility_outlined,
-                        size: 16,
-                        color: theme.colorScheme.onSurfaceVariant),
-                    const SizedBox(width: 6),
-                    // Flexible, or the Row hands the Text unbounded width
-                    // and a long owner name overflows the Wrap on phones.
-                    Flexible(
-                      child: Text(
-                        trip.canEdit
-                            ? (trip.ownerName != null
-                                ? l10n.tripCoPlanningWith(trip.ownerName!)
-                                : l10n.tripCoPlanningShared)
-                            : (trip.ownerName != null
-                                ? l10n.tripSharedBy(trip.ownerName!)
-                                : l10n.tripSharedViewOnly),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: theme.textTheme.bodySmall?.copyWith(
-                            color: theme.colorScheme.onSurfaceVariant),
-                      ),
-                    ),
-                  ],
-                ),
-              if (trip.updatedByName != null)
-                Text(
-                  l10n.tripUpdatedBy(
-                      trip.updatedByName!, _relativeTime(trip.updatedAt)),
-                  style: theme.textTheme.bodySmall
-                      ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-                ),
-            ],
-          ),
-        ],
-        if (overview != null) ...[
-          const SizedBox(height: 12),
-          // Self-contained show-more leaf: toggling it rebuilds this text
-          // block only, not the whole screen.
-          _OverviewText(
-            text: overview,
-            style: theme.textTheme.bodyMedium
-                ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-          ),
-        ],
-      ],
-    );
-  }
-
   // Destination pins live in [TripDerivation.mapDestinations]: one per
   // location group with a real coordinate, in visit order, derived from the
   // FULL itinerary — never the day/category-filtered view, which would shift
   // the numbering (the home-leg doctrine in [_openFullMap]).
-
-  /// The rounded map card shared by the wide (pinned header) and phone
-  /// (scroll-away) layouts. When [expandable], the map is a static preview:
-  /// an [AbsorbPointer] swallows the pin/stay gesture detectors so the whole
-  /// card is one tap target opening [TripMapScreen], and — because the tap
-  /// [GestureDetector] claims only taps — vertical drags fall through to the
-  /// page scroll. The leg chips sit above the gesture layer either way, so
-  /// they stay tappable inline.
-  Widget _mapCard(
-    Trip trip,
-    AppLocalizations l10n, {
-    required bool expandable,
-  }) {
-    final derivation = _derive(trip);
-    final endpoints = derivation.homeLegEndpoints;
-    final legChips = derivation.legChips;
-    // The strip and focus exist only with ≥2 legs — below that the
-    // destination-overview mode never engages, so "All" and "the one leg"
-    // would draw the identical map.
-    final hasChips = legChips.length >= 2;
-    String labelFor(String key) {
-      for (final c in legChips) {
-        if (c.key == key) return c.label;
-      }
-      return key;
-    }
-
-    // Focus + selection live in ValueNotifiers: this builder — not the whole
-    // screen — is what a leg-chip tap or pin/row selection rebuilds. The
-    // leg-filtered lists come from the derivation's lazy per-leg caches, so
-    // a selection-only rebuild passes TripMap the identical lists and its
-    // marker cache skips re-clustering.
-    final card = ListenableBuilder(
-      listenable: Listenable.merge([_focusedLegKey, _selectedPosition]),
-      builder: (context, _) {
-        final focusKey = _clampedLegKey(derivation);
-        // Consumer, not the screen's ref: homeOverlayFor watches the
-        // home-airport resolution provider, and that watch belongs to the
-        // map subtree — resolution landing must not rebuild the screen.
-        Widget map = Consumer(
-          builder: (context, ref, _) {
-            final candidates = homeOverlayFor(
-              ref,
-              homeAirport: _homeAirport,
-              travelMode: trip.travelMode,
-              tripOrigin: trip.origin,
-              tripOriginAirport: trip.originAirport,
-              tripReturnAirport: trip.returnAirport,
-              focusedLegIndex:
-                  focusKey == null ? null : derivation.legIndexOf(focusKey),
-              legCount: derivation.legs.length,
-              firstCityPoint: endpoints.first,
-              lastCityPoint: endpoints.last,
-            );
-            // !expandable == _mapPinned at both _mapCard call sites, so the
-            // wide default rides the same slot-width answer as the layout.
-            final showHome = homeOverlayVisible(
-              choice: ref.watch(homeOverlayChoiceProvider),
-              wideLayout: !expandable,
-            );
-            return TripMap(
-              items: derivation.legFilteredItems(focusKey),
-              accommodations: derivation.legFilteredStays(focusKey),
-              destinations:
-                  focusKey == null ? derivation.mapDestinations : null,
-              selectedPosition: _selectedPosition.value,
-              // Unfiltered by leg: TripMap's position+1 adjacency guard
-              // already keeps labels within a city.
-              segmentLabels: _mapSegmentLabels(derivation),
-              home: showHome ? candidates : const [],
-              homeShown: showHome,
-              // Narrow preview (expandable) deliberately gets no toggle:
-              // interactive:false suppresses the controls anyway, and the
-              // full-screen map — one tap away — carries it.
-              onToggleHome: (expandable || candidates.isEmpty)
-                  ? null
-                  : () => ref
-                      .read(homeOverlayChoiceProvider.notifier)
-                      .setShown(!showHome),
-              fitSignature: focusKey,
-              // Keep fitted markers clear of the chip row overlaid below.
-              topOverlayInset: hasChips ? MapLegChips.mapTopInset : 0,
-              interactive: !expandable,
-              emptyLabel: focusKey == null
-                  ? l10n.tripNoMappedPlaces
-                  : l10n.tripNoPlacesInLeg(labelFor(focusKey)),
-              // The preview absorbs pointers, so its empty-state add-place
-              // hint and CTA would invite an action it can't take (tapping
-              // opens the full-screen map, which has both); dropping them
-              // also keeps the empty state inside the 180px preview card.
-              // The pinned "+ Add place" button sits right below anyway.
-              emptyMessage:
-                  (expandable || _isOffline) ? null : l10n.tripAddPlaceMapHint,
-              emptyAction: (expandable || _isOffline || _readOnly)
-                  ? null
-                  : FilledButton.tonalIcon(
-                      onPressed: () =>
-                          _addPlace(day: derivation.dayForLeg(focusKey)),
-                      icon: const Icon(Icons.add, size: 18),
-                      label: Text(l10n.tripAddPlace),
-                    ),
-              onPinTap: expandable
-                  ? null
-                  : (pos) {
-                      final it =
-                          trip.items!.firstWhere((i) => i.position == pos);
-                      _selectedPosition.value = pos;
-                      // The highlighted row can only be seen if its run
-                      // renders: un-collapse the tapped item's group in
-                      // case the user closed it. Never a focus write —
-                      // that would refit the camera out from under the
-                      // zoom-to-pin move (the invariant holds by
-                      // construction: nothing here touches
-                      // _focusedLegKey). Selection itself is
-                      // notifier-driven, so setState only on a real
-                      // un-collapse.
-                      final d = _derive(trip);
-                      final groupKey =
-                          d.groupKeyForLeg(d.legKeyOfPosition(pos));
-                      if (groupKey != null &&
-                          _collapsedGroups.remove(groupKey)) {
-                        setState(() {});
-                      }
-                      _showSnack(it.name);
-                    },
-              // Region-pin navigation (All overview only — that's the only
-              // mode destination pins render in): scroll the list to the
-              // region's group. Deliberately NO focus write: focusing
-              // would swap the map to per-item pins and delete the very
-              // pins under the pointer; the camera must not move either.
-              onDestinationTap: expandable
-                  ? null // phone preview: AbsorbPointer'd, tap opens full map
-                  : (legKey) {
-                      final d = _derive(trip);
-                      final groupKey = d.groupKeyForLeg(legKey);
-                      if (groupKey == null) return; // lens dropped the leg
-                      if (_collapsedGroups.remove(groupKey)) setState(() {});
-                      // Under a bookings/budget lens no header builds and
-                      // the scroll no-ops harmlessly (the chip-tap stance:
-                      // no lens exit).
-                      _revealCityHeader(groupKey);
-                    },
-              onExpand: expandable ? null : () => _openFullMap(trip),
-            );
-          },
-        );
-        if (expandable) {
-          map = GestureDetector(
-            behavior: HitTestBehavior.opaque,
-            onTap: () => _openFullMap(trip),
-            child: AbsorbPointer(child: map),
-          );
-        }
-        return ClipRRect(
-          borderRadius: AppRadius.lgAll,
-          child: Stack(
-            children: [
-              Positioned.fill(child: map),
-              // Above the map's gesture layer, so chip taps and row scrolls
-              // never pan the map.
-              Positioned(
-                top: 8,
-                left: 8,
-                right: 8,
-                child: MapLegChips(
-                  legs: legChips,
-                  selected: focusKey,
-                  mappedLegKeys: derivation.mappedLegKeys,
-                  onSelected: (k) => _setFocusedLeg(derivation, k),
-                ),
-              ),
-              if (expandable)
-                Positioned(
-                  right: 8,
-                  bottom: 8,
-                  child: MapControlButton(
-                    icon: Icons.fullscreen,
-                    tooltip: l10n.tripExpandMap,
-                    onTap: () => _openFullMap(trip),
-                  ),
-                ),
-            ],
-          ),
-        );
-      },
-    );
-    // Repaint-isolate the card: on the wide layout it sits pinned over the
-    // scrolling itinerary, and without a boundary its picture (card chrome +
-    // chip strip — flutter_map keeps its own internal boundary) is
-    // re-recorded every scroll frame. The boundary also keeps tap-time card
-    // rebuilds from dirtying the page layer. Inside the method so both call
-    // sites (pinned wide band, scroll-away phone card) are covered.
-    return RepaintBoundary(child: card);
-  }
 
   /// Read-side clamp for [_focusedLegKey]: a refresh can remove the focused
   /// leg (or shrink the trip below two legs, where focus is disabled); a key
@@ -6756,9 +3469,26 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.stretch,
                                 children: [
-                                  _buildHeaderCard(trip, theme),
-                                  _nextStepArea(trip),
-                                  _continueChatRow(trip, theme, l10n),
+                                  TripHeaderCard(
+                                    trip: trip,
+                                    narrow: _narrow,
+                                    isOffline: _isOffline,
+                                    readOnly: _readOnly,
+                                    panelOpen: _panelOpen,
+                                    displayTitle: _displayTitle(trip),
+                                    overview: _overviewText(trip),
+                                    onEditDetails: _editTripDetails,
+                                    onEditDates: _editDates,
+                                    onRefine: () => _openRefine(
+                                        trip, const RefineTarget.trip()),
+                                    onNextStepAction: (step) =>
+                                        _onNextStepAction(trip, step),
+                                    onOpenHealthSheet: () =>
+                                        _openHealthSheet(trip),
+                                    transportHandsOff: _transportHandsOff,
+                                    onOpenChat: () => _openChat(trip),
+                                    onNewChat: () => _newChat(trip),
+                                  ),
                                   const SizedBox(height: AppSpacing.lg),
                                 ],
                               ),
@@ -6771,31 +3501,48 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
                           // with the page, and tapping it opens the
                           // full-screen map (TripMapScreen).
                           if (derivation.mapShown)
-                            if (_mapPinned)
-                              SliverPersistentHeader(
-                                pinned: true,
-                                delegate: _PinnedHeaderDelegate(
-                                  height: _mapHeaderHeight,
-                                  backgroundColor:
-                                      theme.scaffoldBackgroundColor,
-                                  padding: EdgeInsets.fromLTRB(
-                                      gutter, 12, gutter, 12),
-                                  child: _mapCard(trip, l10n,
-                                      expandable: false),
-                                ),
-                              )
-                            else
-                              SliverToBoxAdapter(
-                                child: Padding(
-                                  padding: EdgeInsets.fromLTRB(
-                                      gutter, 12, gutter, 12),
-                                  child: SizedBox(
-                                    height: _mapHeightNarrow,
-                                    child: _mapCard(trip, l10n,
-                                        expandable: true),
-                                  ),
-                                ),
+                            tripDetailMapBandSliver(
+                              pinned: _mapPinned,
+                              gutter: gutter,
+                              backgroundColor:
+                                  theme.scaffoldBackgroundColor,
+                              cardBuilder: (expandable) =>
+                                  TripDetailMapBand(
+                                trip: trip,
+                                derivation: derivation,
+                                expandable: expandable,
+                                focusedLegKey: _focusedLegKey,
+                                selectedPosition: _selectedPosition,
+                                homeAirport: _homeAirport,
+                                isOffline: _isOffline,
+                                readOnly: _readOnly,
+                                segmentLabels:
+                                    _mapSegmentLabels(derivation),
+                                onAddPlace: (day) => _addPlace(day: day),
+                                onRevealGroup: (groupKey) {
+                                  if (_collapsedGroups.remove(groupKey)) {
+                                    setState(() {});
+                                  }
+                                },
+                                onRevealCityHeader: _revealCityHeader,
+                                onSetFocusedLeg: (k) =>
+                                    _setFocusedLeg(derivation, k),
+                                onOpenFullMap: () => _openFullMap(trip),
+                                onShowSnack: _showSnack,
+                                clampLegKey: _clampedLegKey,
                               ),
+                              pinnedDelegateBuilder: (
+                                      {required height,
+                                      required backgroundColor,
+                                      required padding,
+                                      required child}) =>
+                                  _PinnedHeaderDelegate(
+                                height: height,
+                                backgroundColor: backgroundColor,
+                                padding: padding,
+                                child: child,
+                              ),
+                            ),
                           // Itinerary/Bookings tab row; pins beneath the map
                           // so the view tabs, Today jump, and the view's add
                           // button stay reachable while scrolling. One
@@ -7428,314 +4175,6 @@ class _TripDetailScreenState extends ConsumerState<TripDetailScreen>
   }
 }
 
-/// Adds or edits a custom booking TODO. A destination (and optional dates)
-/// lets the server build the search link; a pasted link overrides it.
-class _AddBookingTodoDialog extends ConsumerStatefulWidget {
-  final String tripId;
-
-  /// When set, the dialog edits this todo (PATCH) instead of adding one.
-  /// Destination/origin aren't persisted server-side, so they open blank:
-  /// re-entering a destination makes the server rebuild the search link,
-  /// otherwise the existing link is kept.
-  final BookingTodo? existing;
-
-  /// The trip's stated non-flight travel mode ('car'|'train'|'bus'|'ferry'):
-  /// new transport todos then prefer the Rome2Rio link over Google Flights.
-  final String? groundMode;
-
-  const _AddBookingTodoDialog(
-      {required this.tripId, this.existing, this.groundMode});
-
-  @override
-  ConsumerState<_AddBookingTodoDialog> createState() =>
-      _AddBookingTodoDialogState();
-}
-
-class _AddBookingTodoDialogState extends ConsumerState<_AddBookingTodoDialog> {
-  String _kind = 'stay';
-  final _title = TextEditingController();
-  final _destination = TextEditingController();
-  final _origin = TextEditingController();
-  DateTime? _departDate;
-  DateTime? _returnDate;
-  final _url = TextEditingController();
-  bool _saving = false;
-  String? _error;
-
-  @override
-  void initState() {
-    super.initState();
-    final e = widget.existing;
-    if (e != null) {
-      _kind = e.kind;
-      _title.text = e.title;
-      _url.text = e.searchUrl ?? '';
-      _departDate = DateTime.tryParse(e.departDate ?? '');
-      _returnDate = DateTime.tryParse(e.returnDate ?? '');
-    }
-  }
-
-  @override
-  void dispose() {
-    _title.dispose();
-    _destination.dispose();
-    _origin.dispose();
-    _url.dispose();
-    super.dispose();
-  }
-
-  String? _nn(String s) {
-    final t = s.trim();
-    return t.isEmpty ? null : t;
-  }
-
-  String _fmt(DateTime d) =>
-      '${d.year}-${d.month.toString().padLeft(2, '0')}-${d.day.toString().padLeft(2, '0')}';
-
-  Future<void> _pickDate(bool isDepart) async {
-    final now = DateTime.now();
-    final first = now.subtract(const Duration(days: 365));
-    final last = now.add(const Duration(days: 365 * 2));
-    var initial = (isDepart ? _departDate : _returnDate) ?? now;
-    if (initial.isBefore(first)) initial = first;
-    if (initial.isAfter(last)) initial = last;
-    final picked = await showDatePicker(
-      context: context,
-      initialDate: initial,
-      firstDate: first,
-      lastDate: last,
-    );
-    if (picked != null) {
-      setState(() => isDepart ? _departDate = picked : _returnDate = picked);
-    }
-  }
-
-  Future<void> _save() async {
-    final l10n = context.l10n;
-    if (_title.text.trim().isEmpty) {
-      setState(() => _error = l10n.tripTitleRequired);
-      return;
-    }
-    setState(() {
-      _saving = true;
-      _error = null;
-    });
-    try {
-      final isTransport = _kind == 'transport';
-      final isEdit = widget.existing != null;
-      // Note: PATCH is a COALESCE partial update, so an omitted date can't
-      // clear a previously saved one — it just keeps the old value.
-      final body = <String, dynamic>{
-        'kind': _kind,
-        'title': _title.text.trim(),
-        if (_nn(_destination.text) != null)
-          'destination': _nn(_destination.text),
-        if (isTransport && _nn(_origin.text) != null)
-          'origin': _nn(_origin.text),
-        if (_departDate != null) 'depart_date': _fmt(_departDate!),
-        if (!isTransport && _returnDate != null)
-          'return_date': _fmt(_returnDate!),
-        // On edit the field is prefilled with the stored link; sending it
-        // back unchanged would override a destination-driven rebuild (an
-        // explicit search_url wins server-side), so omit it unless the user
-        // actually changed it — COALESCE keeps the stored one.
-        if (_nn(_url.text) != null &&
-            (!isEdit || _nn(_url.text) != widget.existing!.searchUrl))
-          'search_url': _nn(_url.text),
-        // The provider is only a preference for the server's link builder; on
-        // edit, sending it without a destination would overwrite the stored
-        // provider while the old link stays — so only send it when a link is
-        // (re)built from a destination.
-        if (!isEdit || _nn(_destination.text) != null) ...{
-          if (_kind == 'stay') 'provider': 'airbnb',
-          if (isTransport)
-            'provider':
-                widget.groundMode != null ? 'rome2rio' : 'google_flights',
-        },
-        'guests': 1,
-        'passengers': 1,
-      };
-      final svc = ref.read(bookingTodosApiServiceProvider);
-      if (isEdit) {
-        await svc.update(widget.tripId, widget.existing!.id, body);
-      } else {
-        await svc.addTodo(widget.tripId, body);
-      }
-      if (mounted) Navigator.pop(context, true);
-    } catch (e) {
-      setState(() {
-        _saving = false;
-        _error = l10n.tripSaveFailed(friendlyError(l10n, e));
-      });
-    }
-  }
-
-  /// A picker-backed date row: tap to pick, with a clear button when set
-  /// (both dates are optional).
-  Widget _dateField(String label, bool isDepart) {
-    final value = isDepart ? _departDate : _returnDate;
-    return Row(
-      children: [
-        Expanded(
-          child: OutlinedButton.icon(
-            icon: const Icon(Icons.event, size: 18),
-            label: Align(
-              alignment: Alignment.centerLeft,
-              child: Text(value == null ? label : '$label: ${_fmt(value)}'),
-            ),
-            onPressed: () => _pickDate(isDepart),
-          ),
-        ),
-        if (value != null)
-          IconButton(
-            icon: const Icon(Icons.clear, size: 18),
-            tooltip: context.l10n.tripClearDate,
-            onPressed: () => setState(() =>
-                isDepart ? _departDate = null : _returnDate = null),
-          ),
-      ],
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    final isTransport = _kind == 'transport';
-    return AlertDialog(
-      title: Text(widget.existing == null
-          ? l10n.tripAddBooking
-          : l10n.tripEditBooking),
-      content: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            DropdownButtonFormField<String>(
-              initialValue: _kind,
-              decoration: InputDecoration(labelText: l10n.tripFieldType),
-              items: [
-                DropdownMenuItem(
-                    value: 'stay', child: Text(l10n.tripKindStay)),
-                DropdownMenuItem(
-                    value: 'transport', child: Text(l10n.tripKindTransport)),
-                DropdownMenuItem(
-                    value: 'other', child: Text(l10n.tripKindOther)),
-              ],
-              onChanged: (v) => setState(() => _kind = v ?? 'stay'),
-            ),
-            const SizedBox(height: AppSpacing.md),
-            TextField(
-                controller: _title,
-                decoration: InputDecoration(labelText: l10n.tripFieldTitle)),
-            const SizedBox(height: AppSpacing.md),
-            if (isTransport) ...[
-              TextField(
-                  controller: _origin,
-                  decoration:
-                      InputDecoration(labelText: l10n.tripFieldOrigin)),
-              const SizedBox(height: AppSpacing.md),
-            ],
-            TextField(
-                controller: _destination,
-                decoration:
-                    InputDecoration(labelText: l10n.tripFieldDestination)),
-            const SizedBox(height: AppSpacing.md),
-            _dateField(
-                isTransport
-                    ? l10n.tripFieldDepartDate
-                    : l10n.tripFieldCheckIn,
-                true),
-            if (!isTransport) ...[
-              const SizedBox(height: AppSpacing.sm),
-              _dateField(l10n.tripFieldCheckOut, false),
-            ],
-            const SizedBox(height: AppSpacing.md),
-            TextField(
-                controller: _url,
-                decoration: InputDecoration(labelText: l10n.tripFieldLink)),
-            if (_error != null) ...[
-              const SizedBox(height: 8),
-              Text(_error!,
-                  style: TextStyle(color: Theme.of(context).colorScheme.error)),
-            ],
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-            onPressed: () => Navigator.pop(context, false),
-            child: Text(l10n.commonCancel)),
-        FilledButton(
-          onPressed: _saving ? null : _save,
-          child: _saving
-              ? const SizedBox(
-                  height: 18,
-                  width: 18,
-                  child: CircularProgressIndicator(strokeWidth: 2))
-              : Text(l10n.commonSave),
-        ),
-      ],
-    );
-  }
-}
-
-/// Small pill showing a place's part of day (Morning/Afternoon/Evening), tinted
-/// by time so a day's rhythm is scannable at a glance.
-class _TimeOfDayChip extends StatelessWidget {
-  final String timeOfDay;
-
-  /// Icon-only glyph for narrow layouts; the Tooltip carries the label (and
-  /// doubles as the semantics label for screen readers).
-  final bool compact;
-
-  const _TimeOfDayChip({required this.timeOfDay, this.compact = false});
-
-  @override
-  Widget build(BuildContext context) {
-    final icon = switch (timeOfDay) {
-      'morning' => Icons.wb_twilight,
-      'afternoon' => Icons.wb_sunny_outlined,
-      'evening' => Icons.nightlight_outlined,
-      _ => Icons.schedule,
-    };
-    final label = _timeOfDayLabel(context.l10n, timeOfDay);
-    final scheme = Theme.of(context).colorScheme;
-    if (compact) {
-      return Tooltip(
-        message: label,
-        child: Container(
-          padding: const EdgeInsets.all(5),
-          decoration: BoxDecoration(
-            color: scheme.secondaryContainer,
-            borderRadius: BorderRadius.circular(12),
-          ),
-          child: Icon(icon, size: 14, color: scheme.onSecondaryContainer),
-        ),
-      );
-    }
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      decoration: BoxDecoration(
-        color: scheme.secondaryContainer,
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(icon, size: 14, color: scheme.onSecondaryContainer),
-          const SizedBox(width: 4),
-          Text(
-            label,
-            style: Theme.of(context).textTheme.labelSmall?.copyWith(
-                  color: scheme.onSecondaryContainer,
-                  fontWeight: FontWeight.w600,
-                ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
 /// A fixed-height header that scrolls with the page until it reaches the top,
 /// then stays pinned. Used for the trip map and, stacked beneath it, the
 /// itinerary filter bar. The opaque [backgroundColor] fill keeps list content
@@ -7781,158 +4220,6 @@ class _PinnedHeaderDelegate extends SliverPersistentHeaderDelegate {
       oldDelegate.height != height ||
       oldDelegate.backgroundColor != backgroundColor ||
       oldDelegate.padding != padding;
-}
-
-/// Bottom sheet for editing one itinerary item. Returns a map of only the
-/// changed fields (the PATCH endpoint is a partial update), or null/empty on
-/// cancel or no changes.
-class _EditItineraryItemSheet extends StatefulWidget {
-  final ItineraryItem item;
-  const _EditItineraryItemSheet({required this.item});
-
-  @override
-  State<_EditItineraryItemSheet> createState() =>
-      _EditItineraryItemSheetState();
-}
-
-class _EditItineraryItemSheetState extends State<_EditItineraryItemSheet> {
-  late final TextEditingController _name;
-  late final TextEditingController _city;
-  late final TextEditingController _day;
-  String? _category;
-  String? _timeOfDay;
-
-  @override
-  void initState() {
-    super.initState();
-    _name = TextEditingController(text: widget.item.name);
-    _city = TextEditingController(text: widget.item.city ?? '');
-    _day = TextEditingController(text: widget.item.day?.toString() ?? '');
-    _category = widget.item.category;
-    _timeOfDay = widget.item.timeOfDay;
-  }
-
-  @override
-  void dispose() {
-    _name.dispose();
-    _city.dispose();
-    _day.dispose();
-    super.dispose();
-  }
-
-  void _save() {
-    final changes = <String, dynamic>{};
-    final name = _name.text.trim();
-    if (name.isNotEmpty && name != widget.item.name) changes['name'] = name;
-    final city = _city.text.trim();
-    if (city.isNotEmpty && city != (widget.item.city ?? '')) {
-      changes['city'] = city;
-    }
-    final day = int.tryParse(_day.text.trim());
-    if (day != null && day >= 1 && day != widget.item.day) {
-      changes['day'] = day;
-    }
-    if (_category != null && _category != widget.item.category) {
-      changes['category'] = _category;
-    }
-    if (_timeOfDay != null && _timeOfDay != widget.item.timeOfDay) {
-      changes['time_of_day'] = _timeOfDay;
-    }
-    Navigator.of(context).pop(changes);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = context.l10n;
-    return Padding(
-      padding: EdgeInsets.only(
-        left: AppSpacing.lg,
-        right: AppSpacing.lg,
-        top: AppSpacing.lg,
-        bottom: MediaQuery.of(context).viewInsets.bottom + AppSpacing.lg,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(l10n.tripEditPlace, style: theme.textTheme.titleMedium),
-          const SizedBox(height: AppSpacing.md),
-          TextField(
-            controller: _name,
-            decoration: InputDecoration(
-              labelText: l10n.tripFieldName,
-              border: const OutlineInputBorder(),
-            ),
-          ),
-          const SizedBox(height: AppSpacing.md),
-          Row(
-            children: [
-              Expanded(
-                child: TextField(
-                  controller: _city,
-                  decoration: InputDecoration(
-                    labelText: l10n.tripFieldCity,
-                    border: const OutlineInputBorder(),
-                  ),
-                ),
-              ),
-              const SizedBox(width: AppSpacing.md),
-              SizedBox(
-                width: 90,
-                child: TextField(
-                  controller: _day,
-                  keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                    labelText: l10n.tripFieldDay,
-                    border: const OutlineInputBorder(),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: AppSpacing.md),
-          Wrap(
-            spacing: 8,
-            children: [
-              for (final c in const ['attraction', 'restaurant'])
-                ChoiceChip(
-                  label: Text(_categoryLabel(l10n, c)),
-                  selected: _category == c,
-                  onSelected: (sel) =>
-                      setState(() => _category = sel ? c : _category),
-                ),
-            ],
-          ),
-          const SizedBox(height: AppSpacing.sm),
-          Wrap(
-            spacing: 8,
-            children: [
-              for (final t in const ['morning', 'afternoon', 'evening'])
-                ChoiceChip(
-                  label: Text(_timeOfDayLabel(l10n, t)),
-                  selected: _timeOfDay == t,
-                  onSelected: (sel) =>
-                      setState(() => _timeOfDay = sel ? t : _timeOfDay),
-                ),
-            ],
-          ),
-          const SizedBox(height: AppSpacing.lg),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              TextButton(
-                onPressed: () => Navigator.of(context).pop(),
-                child: Text(l10n.commonCancel),
-              ),
-              const SizedBox(width: AppSpacing.sm),
-              FilledButton(onPressed: _save, child: Text(l10n.commonSave)),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
 }
 
 /// Owner-only bottom sheet: lists active co-planners with per-person removal.
@@ -8151,188 +4438,5 @@ class _CoPlannersSheetState extends ConsumerState<_CoPlannersSheet> {
         ),
       ),
     );
-  }
-}
-
-/// Bottom sheet with a drag-and-drop list for one itinerary section. Pops
-/// with the reordered items on Save, or null on dismiss.
-class _ReorderSectionSheet extends StatefulWidget {
-  final List<ItineraryItem> items;
-  const _ReorderSectionSheet({required this.items});
-
-  @override
-  State<_ReorderSectionSheet> createState() => _ReorderSectionSheetState();
-}
-
-class _ReorderSectionSheetState extends State<_ReorderSectionSheet> {
-  late List<ItineraryItem> _items;
-
-  @override
-  void initState() {
-    super.initState();
-    _items = List.of(widget.items);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = context.l10n;
-    return Padding(
-      padding: const EdgeInsets.all(AppSpacing.lg),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(l10n.tripReorderPlaces, style: theme.textTheme.titleLarge),
-          const SizedBox(height: AppSpacing.xs),
-          Text(
-            l10n.tripReorderHint,
-            style: theme.textTheme.bodySmall
-                ?.copyWith(color: theme.colorScheme.onSurfaceVariant),
-          ),
-          const SizedBox(height: AppSpacing.md),
-          Flexible(
-            child: ReorderableListView.builder(
-              shrinkWrap: true,
-              buildDefaultDragHandles: false,
-              itemCount: _items.length,
-              onReorder: (oldIndex, newIndex) {
-                setState(() {
-                  if (newIndex > oldIndex) newIndex--;
-                  final moved = _items.removeAt(oldIndex);
-                  _items.insert(newIndex, moved);
-                });
-              },
-              itemBuilder: (context, i) {
-                final item = _items[i];
-                return ListTile(
-                  key: ValueKey(item.id),
-                  dense: true,
-                  leading: Text('${i + 1}',
-                      style: theme.textTheme.labelLarge?.copyWith(
-                          color: theme.colorScheme.onSurfaceVariant)),
-                  title: Text(item.name,
-                      maxLines: 1, overflow: TextOverflow.ellipsis),
-                  trailing: ReorderableDragStartListener(
-                    index: i,
-                    child: const Icon(Icons.drag_indicator),
-                  ),
-                );
-              },
-            ),
-          ),
-          const SizedBox(height: AppSpacing.md),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              TextButton(
-                onPressed: () => Navigator.of(context).pop(),
-                child: Text(l10n.commonCancel),
-              ),
-              const SizedBox(width: AppSpacing.sm),
-              FilledButton(
-                onPressed: () => Navigator.of(context).pop(_items),
-                child: Text(l10n.tripSaveOrder),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// The trip overview paragraph with its "Show more/less" toggle — a
-/// self-contained leaf so expanding it rebuilds this block, never the
-/// screen that hosts it. Collapsed = [_OverviewTextState._collapsedMaxLines]
-/// lines; the toggle renders only when that clamp actually clips at the
-/// laid-out width (a character-count heuristic misfired both ways: wide
-/// windows got a dead toggle, short newline-heavy text got none).
-class _OverviewText extends StatefulWidget {
-  final String text;
-  final TextStyle? style;
-
-  const _OverviewText({required this.text, this.style});
-
-  @override
-  State<_OverviewText> createState() => _OverviewTextState();
-}
-
-class _OverviewTextState extends State<_OverviewText> {
-  /// One constant read by BOTH the rendered Text.maxLines and the measuring
-  /// painter, so the render and the toggle decision cannot drift.
-  static const int _collapsedMaxLines = 2;
-
-  bool _expanded = false;
-
-  /// Whether the collapsed clamp would clip the text at [maxWidth] — the
-  /// same verdict the collapsed Text's RenderParagraph reaches. Mirrors
-  /// Text.build: DefaultTextStyle merge for inherited styles, the boldText
-  /// accessibility merge (Text applies it internally; TextPainter does not),
-  /// the ambient TextScaler OBJECT (Android 14+ scaling is nonlinear),
-  /// directionality, locale, and the same ellipsis. Same measurement pattern
-  /// as [_dateChipWidth].
-  bool _collapsedClips(BuildContext context, double maxWidth) {
-    var style = widget.style;
-    if (style == null || style.inherit) {
-      style = DefaultTextStyle.of(context).style.merge(style);
-    }
-    if (MediaQuery.boldTextOf(context)) {
-      style = style.copyWith(fontWeight: FontWeight.bold);
-    }
-    final tp = TextPainter(
-      text: TextSpan(text: widget.text, style: style),
-      maxLines: _collapsedMaxLines,
-      ellipsis: '…',
-      textDirection: Directionality.of(context),
-      textScaler: MediaQuery.textScalerOf(context),
-      locale: Localizations.maybeLocaleOf(context),
-    )..layout(maxWidth: maxWidth);
-    final clips = tp.didExceedMaxLines;
-    tp.dispose();
-    return clips;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    // LayoutBuilder so the toggle decision sees the exact width the Text
-    // wraps at. Synchronous, one layout pass — no post-frame re-measure, so
-    // the header extent is settled the frame it builds (the today-mode
-    // auto-scroll in the hosting scroll view assumes settled extents).
-    return LayoutBuilder(builder: (context, constraints) {
-      // Unbounded width can't wrap, so it can't clip (and never occurs
-      // under this stretched header column).
-      final clips = constraints.maxWidth.isFinite &&
-          _collapsedClips(context, constraints.maxWidth);
-      return Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Text(
-            widget.text,
-            style: widget.style,
-            maxLines: _expanded ? null : _collapsedMaxLines,
-            overflow: _expanded ? TextOverflow.visible : TextOverflow.ellipsis,
-          ),
-          // _expanded is deliberately not reset when the toggle disappears
-          // (window grown until the text fits): expanded and collapsed
-          // renders of fitting text are pixel-identical, so the stale flag
-          // is unobservable.
-          if (clips)
-            Align(
-              alignment: Alignment.centerLeft,
-              child: TextButton(
-                style: TextButton.styleFrom(
-                  padding: EdgeInsets.zero,
-                  minimumSize: const Size(0, 32),
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () => setState(() => _expanded = !_expanded),
-                child: Text(_expanded ? l10n.tripShowLess : l10n.tripShowMore),
-              ),
-            ),
-        ],
-      );
-    });
   }
 }
